Analyze Blocks

Metacells.AnalyzeBlocks — Module

Do simple blocks analysis.

Neighborhoods

Metacells.AnalyzeBlocks.compute_matrix_of_is_in_neighborhood_per_block_per_block! — Function

compute_matrix_of_is_in_neighborhood_per_block_per_block!(
    daf::DafWriter;
    min_neighbour_confusion_fractions::AbstractFloat = ```0.01```,
    min_blocks_in_neighborhood::Integer = ```5```,
    min_metacells_in_neighborhood::Integer = ```20```,
    min_total_UMIs_in_neighborhood::Integer = ```2000000```,
    max_blocks_in_neighborhood::Integer = ```100```,
    max_metacells_in_neighborhood::Integer = ```400```
    max_total_UMIs_in_neighborhood::Integer = ```40000000```,
    overwrite::Bool = false,
)::Nothing

Compute and set matrix_of_is_in_neighborhood_per_block_per_block .

We normalize the confusion matrix between blocks so that each column sums to 1 (that is, we look at the fraction of the cells of each block that are confused with each other block).

We then make it symmetric by adding it to its transpose.

For each block, we sort all blocks by the normalized symmetric confusion score. To break ties (especially for zero confusion), we use the mean distance between the blocks. We then pick the first few blocks to be in the neighborhood.

All blocks that have a symmetric normalized confusion fraction of at least min_neighbour_confusion_fractions are included, but we stop if we exceed max_blocks_in_neighborhood , max_metacells_in_neighborhood , or max_total_UMIs_in_neighborhood .

If we didn't reach min_blocks_in_neighborhood , min_metacells_in_neighborhood or min_total_UMIs_in_neighborhood , we keep adding blocks even if their symmetric normalized confusion fraction is low (even zero).

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each block.

block @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each block.

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

Matrices

block, block @ confusion by closest by pertinent_markers ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells which are included in each column (block) and are closest to each (row) block.

block, block @ mean euclidean skeleton fold distance ::Union{Float32, Float64} (required): The mean Euclidean skeleton genes fractions distance between the metacells of the blocks.

Outputs

Matrices

block, block @ is in neighborhood ::Bool (created): For each (column) block, whether each (row) block is in its immediate neighborhood.

Counts

Metacells.AnalyzeBlocks.compute_vector_of_n_cells_per_block! — Function

function compute_vector_of_n_cells_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_cells_per_block .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each metacell.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Outputs

Vectors

block @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of cells in each block.

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_cells_per_block! — Function

function compute_vector_of_n_neighborhood_cells_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of cells in the metacells of the blocks of the neighborhood centered at a block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each block.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of cells in the metacells in each block's neighborhood.

Metacells.AnalyzeBlocks.compute_vector_of_n_metacells_per_block! — Function

function compute_vector_of_n_metacells_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_metacells_per_block .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Outputs

Vectors

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of metacells in each block.

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_metacells_per_block! — Function

function compute_vector_of_n_neighborhood_metacells_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_neighborhood_metacells_per_block

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood metacells ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of metacells in the blocks in each block's neighborhood.

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_blocks_per_block! — Function

function compute_vector_of_n_neighborhood_blocks_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_neighborhood_blocks_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood blocks ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of (rows) blocks in each (column) block's neighborhood.

UMIs

Metacells.AnalyzeBlocks.compute_matrix_of_UMIs_per_gene_per_block! — Function

function compute_matrix_of_UMIs_per_gene_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_UMIs_per_gene_per_block .

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of UMIs of each gene in each block.

Metacells.AnalyzeBlocks.compute_vector_of_total_UMIs_per_block! — Function

function compute_vector_of_total_UMIs_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of non-excluded genes per block.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Vectors

block @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of UMIs of all the non-excluded genes in each block.

Metacells.AnalyzeBlocks.compute_vector_of_total_neighborhood_UMIs_per_block! — Function

function compute_vector_of_total_neighborhood_UMIs_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_total_neighborhood_UMIs_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each block.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Outputs

Vectors

block @ total neighborhood UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of non-excluded gene UMIs in the cells in the metacells in each block's neighborhood.

Metacells.AnalyzeBlocks.compute_matrix_of_linear_fraction_per_gene_per_block! — Function

function compute_matrix_of_linear_fraction_per_gene_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_linear_fraction_per_gene_per_block .

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

block @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each block.

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Matrices

gene, block @ linear_fraction ::Union{Float32, Float64} (created): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Metacells.AnalyzeBlocks.compute_matrix_of_log_linear_fraction_per_gene_per_block! — Function

function compute_matrix_of_log_linear_fraction_per_gene_per_block!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_log_linear_fraction_per_gene_per_block . This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

gene, block @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Outputs

Matrices

gene, block @ log linear fraction ::Union{Float32, Float64} (created): The log base 2 of the linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Genes

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_marker_per_gene_per_block! — Function

compute_matrix_of_is_neighborhood_marker_per_gene_per_block!(
    daf::DafWriter;
    min_marker_gene_max_fraction::AbstractFloat = 2 ^ -13.5,
    min_marker_gene_range_fold::Real = 1.0,
    min_marker_quantile::Real = 0.1,
    overwrite::Bool = false,
)::Nothing

Compute and set matrix_of_is_neighborhood_marker_per_gene_per_block .

We only consider genes which are markers in the overall population. We then call compute_vector_of_is_marker_per_gene! but only looking at the metacells in each neighborhood.

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_distinct_per_gene_per_block! — Function

compute_matrix_of_is_neighborhood_distinct_per_gene_per_block!(
    daf::DafWriter;
    min_distinct_gene_max_fraction::AbstractFloat = 2 ^ -14.5,
    min_distinct_gene_mean_fold::Real = 2,
    overwrite::Bool = false,
)::Nothing

Compute and set matrix_of_is_in_neighborhood_per_block_per_block . A gene is distinct in the neighborhood if:

It has a maximal expression level of at least min_distinct_gene_max_fraction .

When comparing its fold factor compared to its median expression level in the population, then the mean fold factor in the neighborhood is at least min_distinct_gene_mean_fold . This is not the absolute fold, that is, we only look for genes which are stronger than the population as a whole, not weaker.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Matrices

block, gene @ is neighborhood distinct ::Bool (created): A mask of genes that distinguish between cell states of each block's neighborhood and the rest of the manifold.

Metacells.AnalyzeBlocks.compute_matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block! — Function

compute_matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block!(
    daf::DafWriter;
    min_gene_correlation::AbstractFloat = function_default(
        compute_vector_of_is_correlated_with_skeleton_per_gene!,
        :min_gene_correlation,
    ),
    min_gene_correlation_quantile::AbstractFloat = function_default(
        compute_vector_of_is_correlated_with_skeleton_per_gene!,
        :min_gene_correlation_quantile,
    ),
    genes_correlation_window::Integer = function_default(
        compute_vector_of_is_correlated_with_skeleton_per_gene!,
        :genes_correlation_window,
    ),
    overwrite::Bool = false,
)::Nothing

Compute and set matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block . This just invokes compute_vector_of_is_correlated_with_skeleton_per_gene! for the metacells in each neighborhood.

Metacells.AnalyzeBlocks.compute_matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block! — Function

compute_matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block . This takes a long time. Still better than doing full two-way cross validation, though, and only needed to be done once (for some base metacells).

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of cells in the metacells in each block's neighborhood.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, block @ is neighborhood marker ::Bool (required): A mask of genes that distinguish between cell states in each block's neighborhood.

Outputs

Matrices

gene, block @ most correlated gene in neighborhood ::AbstractString (created): The most correlated neighborhood marker gene of each (base) neighborhood marker gene per block.

Types

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_metacells! — Function

compute_vector_of_type_per_block_by_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_block by using the type of the metacells grouped into each block. The most frequent ("mode") of the type is used for each block. This method is used when metacells are annotated with types (which is typically a supervised process).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ type ::AbstractString (required): The type each metacell belongs to.

Outputs

Vectors

block @ type ::AbstractString (created): The type each block belongs to.

Metacells.AnalyzeBlocks.compute_vector_of_type_per_metacell_by_blocks! — Function

compute_vector_of_type_per_metacell_by_blocks!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_metacell by using the type of the blocks the metacells belong to. This method is used when blocks are annotated with types (which is typically a supervised process).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ type ::AbstractString (required): The type each block belongs to.

Outputs

Vectors

metacell @ type ::AbstractString (created): The type each metacell belongs to.

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_cells! — Function

compute_vector_of_type_per_block_by_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_block by using the type of the metacells grouped into each block. The most frequent ("mode") of the type is used for each block. This method is used when cells are annotated with types (typically when importing single-cell data with type annotations).

Inputs

Axes

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ type ::AbstractString (required): The type each metacell belongs to.

Outputs

Vectors

block @ type ::AbstractString (created): The type each block belongs to.

Metacells.AnalyzeBlocks.compute_vector_of_type_per_cell_by_blocks! — Function

compute_vector_of_type_per_cell_by_blocks!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_cell by using the type of the blocks the metacells of the cells belong to. This method is used when blocks are annotated with types (which is typically a supervised process).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ type ::AbstractString (required): The type each block belongs to.

Outputs

Vectors

metacell @ type ::AbstractString (created): The type each metacell belongs to.

Distances

Metacells.AnalyzeBlocks.compute_matrix_of_mean_euclidean_skeleton_fold_distance_between_blocks! — Function

function compute_matrix_of_mean_euclidean_skeleton_fold_distance_between_blocks!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The maximal significant skeleton genes fractions fold factor between metacells of the blocks.

Metacells.AnalyzeBlocks.compute_vector_of_block_closest_by_pertinent_markers_per_cell! — Function

compute_vector_of_block_closest_by_pertinent_markers_per_cell!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool = false
)::Nothing

Compute and set vector_of_block_closest_by_pertinent_markers_per_cell .

Metacells.AnalyzeBlocks.compute_matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block! — Function

compute_matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block!(daf::DafWriter; overwrite::Bool = false)::Nothing

Compute and set matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block .

Inputs

Axes

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

cell @ block.closest by pertinent_markers ::AbstractString (required): The block each cell is closest to considering only the non-lateral marker genes.

Outputs

Matrices

block, block @ confusion by closest by pertinent_markers ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of cells which are included in each column (block) and are closest to each (row) block.

Correlations

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block! — Function

compute_matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool,
)::Nothing

Compute and set matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of cells in the metacells in each block's neighborhood.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

gene, block @ is neighborhood marker ::Bool (required): A mask of genes that distinguish between cell states in each block's neighborhood.

Outputs

Matrices

gene, block @ correlation between neighborhood cells and punctuated metacells ::Union{Float32, Float64} (created): The correlation between cells and their metacells (minus the correlated cell) gene expression levels in each block's neighborhood.

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block! — Function

compute_matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block!(;
    other_daf::DafWriter,
    base_daf::DafReader,
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set [matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block ](@ref). This will also copy base_block_axis from the base_daf into the other_daf if needed.

Other

Inputs

Axes

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, base block @ correlation between base neighborhood cells and punctuated metacells ::Union{Float32, Float64} (created): The correlation between cells and their metacells (minus the correlated cell) of each gene's expression levels in each base block's neighborhood.

Base

Inputs

Axes

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of cells in the metacells in each block's neighborhood.

Matrices

gene, block @ is neighborhood marker ::Bool (required): A mask of genes that distinguish between cell states in each block's neighborhood.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_with_most_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block! — Function

compute_matrix_of_correlation_with_most_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block!(;
    other_daf::DafWriter,
    base_daf::DafReader;
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_correlation_with_most_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block . This will also copy base_block_axis from the base_daf into the other_daf if needed.

The cell and gene axes must be identical in both base_daf and other_daf .

That is, we evaluate the quality of the other_daf metacells by correlating the genes in each cell in each neighborhood of base_daf with the expression of their most correlated gene in their other_daf punctuated metacells (metacells minus the correlated cell). This works as a poor man's cross-validation; not as good as real cross-validation but much less expensive to compute.

Other

Inputs

Axes

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, base block @ correlation with most between base neighborhood cells and punctuated metacells ::Union{Float32, Float64} (created): The correlation of each gene with the most correlated other gene between the cells of a base neighborhood and their punctuated metacells of each base neighborhood.

Base

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of cells in the metacells in each block's neighborhood.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, block @ most correlated gene in neighborhood ::AbstractString (required): The most correlated neighborhood marker gene of each (base) neighborhood marker gene per block.

Additional inputs may be used depending on the parameter(s).

Additional outputs may be created depending on the parameter(s).

UMAP

Metacells.AnalyzeBlocks.compute_blocks_2d_umap_by_metacells! — Function

function compute_blocks_2d_umap_by_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_umap_x_per_block and vector_of_umap_y_per_block by taking the mean of the 2D UMAP coordinates of the metacells in each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ umap_x ::Union{Float32, Float64} (required): The X coordinate of each metacell in a 2D UMAP projection.

metacell @ umap_y ::Union{Float32, Float64} (required): The Y coordinate of each metacell in a 2D UMAP projection.

Outputs

Vectors

block @ umap_x ::Union{Float32, Float64} (created): The X coordinate of each block in a 2D UMAP projection.

block @ umap_y ::Union{Float32, Float64} (created): The Y coordinate of each block in a 2D UMAP projection.

Metacells.AnalyzeBlocks.compute_blocks_3d_umap_by_metacells! — Function

function compute_blocks_3d_umap_by_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_umap_u_per_block , vector_of_umap_v_per_block and vector_of_umap_w_per_block by taking the mean of the 3D UMAP coordinates of the metacells in each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ umap_u ::Union{Float32, Float64} (required): The U coordinate of each metacell in a 3D UMAP projection.

metacell @ umap_v ::Union{Float32, Float64} (required): The V coordinate of each metacell in a 3D UMAP projection.

metacell @ umap_w ::Union{Float32, Float64} (required): The W coordinate of each metacell in a 3D UMAP projection.

Outputs

Vectors

block @ umap_u ::Union{Float32, Float64} (created): The U coordinate of each block in a 3D UMAP projection.

block @ umap_v ::Union{Float32, Float64} (created): The V coordinate of each block in a 3D UMAP projection.

block @ umap_w ::Union{Float32, Float64} (created): The W coordinate of each block in a 3D UMAP projection.

Index

Metacells.AnalyzeBlocks

Metacells.AnalyzeBlocks.compute_blocks_2d_umap_by_metacells!

Metacells.AnalyzeBlocks.compute_blocks_3d_umap_by_metacells!

Metacells.AnalyzeBlocks.compute_matrix_of_UMIs_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_with_most_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_in_neighborhood_per_block_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_distinct_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_marker_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_linear_fraction_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_log_linear_fraction_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_mean_euclidean_skeleton_fold_distance_between_blocks!

Metacells.AnalyzeBlocks.compute_matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_block_closest_by_pertinent_markers_per_cell!

Metacells.AnalyzeBlocks.compute_vector_of_n_cells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_metacells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_blocks_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_cells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_metacells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_total_UMIs_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_total_neighborhood_UMIs_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_cells!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_metacells!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_cell_by_blocks!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_metacell_by_blocks!