Analyze Cells

Metacells.AnalyzeCells — Module

Do simple per-cell analysis.

Metacells.AnalyzeCells.compute_cells_total_UMIs! — Function

function compute_cells_total_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute the total UMIs of the genes in each cell.

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of all the non-excluded genes in each cell.

Metacells.AnalyzeCells.compute_cells_covered_UMIs! — Function

function compute_cells_covered_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute the total UMIs of the covered genes in each cell.

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

cell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of all the covered genes in each cell.

Index

Metacells.AnalyzeCells

Metacells.AnalyzeCells.compute_cells_covered_UMIs!

Metacells.AnalyzeCells.compute_cells_total_UMIs!