Analyze Genes

Metacells.AnalyzeGenes — Module

Do simple per-gene analysis.

Metacells.AnalyzeGenes.compute_genes_divergence! — Function

function compute_genes_divergence!(
    daf::DafWriter;
    min_divergent_gene_range_fold::Real = ```7```,
    overwrite::Bool = ```false```,
)::Nothing

Compute a divergence factor for all genes. We need this because when computing any form of squared error, "bursty" genes with a very wide range of expression level tend to over-influence the results. We therefore compute a scaled expression level by multiplying the raw expression level by 1.0 - divergence . For most ("well behaved") genes, the divergence will be 0. For genes whose fold factor between the highest and lowest (log base 2) expression level is above min_divergent_gene_range_fold , we compute a divergence factor such that the above scaling will reduce this fold factor to the min_divergent_gene_range_fold threshold.

Note

The computation depends on the gene_fraction_regularization , so it is important to give it the same value as the one given to compute_metacells_log_*_fractions! .

This is admittedly a blunt instrument, but it is the least-bad option we have for now. It beats the old method of manually constructing an is_noisy genes mask.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The estimated linear fraction of the UMIs of each gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ divergence ::Union{Float32, Float64} (guaranteed): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

Metacells.AnalyzeGenes.identify_marker_genes! — Function

function identify_marker_genes!(
    daf::DafWriter;
    min_marker_gene_max_fraction::AbstractFloat = ```0.0001```,
    min_marker_gene_range_fold::Real = ```2```,
    overwrite::Bool = ```false```,
)::Nothing

Identify the genes that distinguish at least one metacell from the rest. Such genes are called "marker" genes as they (potentially) mark specific cell states. If overwrite , will overwrite an existing is_marker mask.

Marker genes are genes which:

Have a maximal gene expression level of at least min_marker_gene_max_fraction , and

Have a range of expression of at least min_marker_gene_range_fold between the maximal and minimal (log base 2) of the expression level.

Note

This uses the virtual metacell_gene_fraction_matrix and metacell_gene_log_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...). Make sure you are consistent when mapping the fractions and log-fraction matrices.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): Some estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ log_fraction ::Union{Float32, Float64} (required): The log base 2 of some estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

gene @ is_marker ::Bool (guaranteed): A mask of genes that distinguish between cell states.

Metacells.AnalyzeGenes.rank_marker_genes! — Function

function rank_marker_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute the relative ranks of marker genes.

Compute the median of the (log base 2) of the gene expression level for each marker gene across all the metacells.

Compute the per-marker-per-metacell fold factor (absolute difference of the log expression from the median).

Rank the markers for each metacell (1 having the largest fold factor relative to the median).

For each marker, give it a priority which is a tuple of (1) the minimal rank it has in all metacells (2) the maximal fold it has in metacells where it has that rank (negated).

Sort the markers according to this priority.

Non-marker genes are given a rank of typemax(UInt32) regardless of their expression level in the metacells.

Note

This uses the virtual metacell_gene_log_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

Matrices

gene, metacell @ log_fraction ::Union{Float32, Float64} (required): The log base 2 of some estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

gene @ marker_rank ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The ralative ranks of the marker genes.

Metacells.AnalyzeGenes.identify_uncorrelated_genes! — Function

function identify_uncorrelated_genes!(
    daf::DafWriter;
    correlation_confidence::AbstractFloat = ```0.99```,
    overwrite::Bool = ```false```,
)::Nothing

Identify marker genes that are correlated with other gene(s). Such genes are good candidates for looking for groups of genes that act together. If overwrite , will overwrite an existing is_correlated mask.

Correlate the (log base 2 of the) expression level between all the pairs of genes.

For each gene, shuffle its values along all metacells, and again correlate this between all the pairs of genes.

Find the maximal absolute correlation for each gene in both cases (that is, strong anti-correlation also counts).

Find the correlation_confidence quantile correlation of the shuffled data.

Identify the genes that have at least that level of correlations in the unshuffled data.

Note

This uses the virtual metacell_gene_log_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...). While at it, you should probably also restrict the set of genes to the marker genes.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ log_fraction ::Union{Float32, Float64} (required): The log base 2 of some estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

gene @ is_uncorrelated ::Bool (guaranteed): A mask of genes that are not correlated with other gene(s).

Metacells.AnalyzeGenes.identify_covered_genes! — Function

function identify_covered_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Identify the genes that will be approximated by the local linear programs. Picking them is simple: we cover all marker genes, that are not lateral, and not uncorrelated.

Inputs

Axes

gene (required): Sequenced genes.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

gene @ is_uncorrelated ::Bool (required): A mask of genes that are not correlated with other gene(s).

Outputs

Vectors

gene @ is_covered ::Bool (guaranteed): A mask of genes that are covered by the local linear program.

Metacells.AnalyzeGenes.identify_skeleton_genes! — Function

function identify_skeleton_genes!(
    daf::DafWriter;
    max_skeleton_genes::Integer = ```200```,
    overwrite::Bool = ```false```,
)::Nothing

Identify the skeleton genes that will be used to predict the rest of the (covered) genes. We just pick the max_skeleton_genes that have the lowest marker_rank out of the covered genes.

Inputs

Axes

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

gene @ marker_rank ::Union{UInt16, UInt32, UInt64, UInt8} (required): The ralative ranks of the marker genes.

Outputs

Vectors

gene @ is_skeleton ::Bool (guaranteed): A mask of genes that are used to predict the values of the rest of the (covered) genes.

Index

Metacells.AnalyzeGenes

Metacells.AnalyzeGenes.compute_genes_divergence!

Metacells.AnalyzeGenes.identify_covered_genes!

Metacells.AnalyzeGenes.identify_marker_genes!

Metacells.AnalyzeGenes.identify_skeleton_genes!

Metacells.AnalyzeGenes.identify_uncorrelated_genes!

Metacells.AnalyzeGenes.rank_marker_genes!