Analyze Metacells

Metacells.AnalyzeMetacells — Module

Do simple metacells analysis.

Counts

Metacells.AnalyzeMetacells.compute_vector_of_n_cells_per_metacell! — Function

function compute_vector_of_n_cells_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_cells_per_metacell .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Outputs

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of cells in each metacell.

UMIs

Metacells.AnalyzeMetacells.compute_matrix_of_UMIs_per_gene_per_metacell! — Function

function compute_matrix_of_UMIs_per_gene_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_UMIs_per_gene_per_metacell .

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of UMIs of each gene in the cells of each metacell.

Metacells.AnalyzeMetacells.compute_vector_of_total_UMIs_per_metacell! — Function

function compute_vector_of_total_UMIs_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_total_UMIs_per_metacell .

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (created): The total number of UMIs of all the non-excluded genes in each metacell.

Metacells.AnalyzeMetacells.compute_matrix_of_linear_fraction_per_gene_per_metacell! — Function

function compute_matrix_of_linear_fraction_per_gene_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_linear_fraction_per_gene_per_metacell .

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (created): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Metacells.AnalyzeMetacells.compute_matrix_of_log_linear_fraction_per_gene_per_metacell! — Function

function compute_matrix_of_log_linear_fraction_per_gene_per_metacell!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_log_linear_fraction_per_gene_per_metacell . This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (created): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Types

Metacells.AnalyzeMetacells.compute_vector_of_type_per_cell_by_metacells! — Function

function compute_vector_of_type_per_cell_by_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_cell by using the type of the metacell each cell belongs to. Cells that do not belong to any metacells (including excluded cells) are given an empty type. This method is used when metacells are annotated with types (which is typically a supervised process).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ type ::AbstractString (required): The type each metacell belongs to.

Outputs

Vectors

cell @ type ::AbstractString (created): The type each cell belongs to.

Metacells.AnalyzeMetacells.compute_vector_of_type_per_metacell_by_cells! — Function

function compute_vector_of_type_per_metacell_by_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_type_per_metacell by using the type of the cells grouped into each metacell. The most frequent ("mode") of the type is used for each metacell. This method is used when cells are annotated with types (typically when importing single-cell data with type annotations).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

cell @ type ::AbstractString (required): The type each cell belongs to.

Outputs

Vectors

metacell @ type ::AbstractString (created): The type each metacell belongs to.

Distances

Metacells.AnalyzeMetacells.compute_matrix_of_euclidean_skeleton_fold_distance_between_metacells! — Function

function compute_matrix_of_euclidean_skeleton_fold_distance_between_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_euclidean_skeleton_fold_distance_between_metacells .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the genes.

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Matrices

metacell, metacell @ euclidean skeleton fold_distance ::Union{Float32, Float64} (created): The Euclidean distance between the log of the fraction of the skeleton genes between the metacells.

Metacells.AnalyzeMetacells.compute_matrix_of_max_skeleton_fold_distance_between_metacells! — Function

function compute_matrix_of_max_skeleton_fold_distance_between_metacells!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    min_significant_gene_UMIs::Integer = ```40```,
    fold_confidence::AbstractFloat = ```0.9```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_max_skeleton_fold_distance_between_metacells . For computing this fold factor, we add the gene_fraction_regularization to deal with zero fractions. We ignore "insignificant" genes whose total UMIs in the compared metacells isn't at least min_significant_gene_UMIs . We also we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells. That is, when we have a high fold factor, we are pretty certain the metacells truly have a significant difference in the expression of the skeleton gene.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the genes.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

metacell, metacell @ max skeleton fold_distance ::Union{Float32, Float64} (created): The maximal significant fold factor between the fraction of skeleton genes between the metacells.

Correlations

Metacells.AnalyzeMetacells.compute_vector_of_correlation_between_cells_and_punctuated_metacells_per_gene! — Function

function compute_vector_of_correlation_between_cells_and_punctuated_metacells_per_gene!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = (DEFAULT.gene_fraction_regularization),
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_correlation_between_cells_and_punctuated_metacells_per_gene .

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Vectors

gene @ correlation between cells and punctuated_metacells ::Union{Float32, Float64} (created): The correlation between cells and their metacells (minus the correlated cell) gene expression levels.

Metacells.AnalyzeMetacells.compute_matrix_of_correlation_per_gene_per_gene_of_subset_of_metacells! — Function

compute_matrix_of_correlation_per_gene_per_gene_of_subset_of_metacells!(
    daf::DafWriter;
    metacells_subset::Union{QueryString, BitVector, AbstractVector{Bool}},
    matrix_name::AbstractString,
    overwrite::Bool = false,
)::Nothing

Correlate the marker genes in an arbitrary metacells_subset . This subset can be either an explicit mask or a query that masks the metacell axis. The result is stored in a per-gene-per-gene matrix_name .

UMAP

Metacells.AnalyzeMetacells.compute_metacells_2d_umap! — Function

function compute_metacells_2d_umap!(
    daf::DafWriter;
    min_dist::AbstractFloat = ```0.5```,
    n_neighbors::Integer = ```15```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_umap_x_per_metacell and vector_of_umap_y_per_metacell by computing a 2D UMAP projection from the Euclidean skeleton fold distance between metacells.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

metacell, metacell @ euclidean skeleton fold_distance ::Union{Float32, Float64} (required): The Euclidean distance between the log of the fraction of the skeleton genes between the metacells.

Outputs

Vectors

metacell @ umap_x ::Union{Float32, Float64} (created): The X coordinate of each metacell in a 2D UMAP projection.

metacell @ umap_y ::Union{Float32, Float64} (created): The Y coordinate of each metacell in a 2D UMAP projection.

Metacells.AnalyzeMetacells.compute_metacells_3d_umap! — Function

function compute_metacells_3d_umap!(
    daf::DafWriter;
    min_dist::AbstractFloat = ```0.5```,
    n_neighbors::Integer = ```15```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_umap_u_per_metacell , vector_of_umap_v_per_metacell and vector_of_umap_w_per_metacell by computing a 3D UMAP projection from the Euclidean skeleton fold distance between metacells.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

metacell, metacell @ euclidean skeleton fold_distance ::Union{Float32, Float64} (required): The Euclidean distance between the log of the fraction of the skeleton genes between the metacells.

Outputs

Vectors

metacell @ umap_u ::Union{Float32, Float64} (created): The U coordinate of each metacell in a 3D UMAP projection.

metacell @ umap_v ::Union{Float32, Float64} (created): The V coordinate of each metacell in a 3D UMAP projection.

metacell @ umap_w ::Union{Float32, Float64} (created): The W coordinate of each metacell in a 3D UMAP projection.

Index

Metacells.AnalyzeMetacells

Metacells.AnalyzeMetacells.compute_matrix_of_UMIs_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_correlation_per_gene_per_gene_of_subset_of_metacells!

Metacells.AnalyzeMetacells.compute_matrix_of_euclidean_skeleton_fold_distance_between_metacells!

Metacells.AnalyzeMetacells.compute_matrix_of_linear_fraction_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_log_linear_fraction_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_max_skeleton_fold_distance_between_metacells!

Metacells.AnalyzeMetacells.compute_metacells_2d_umap!

Metacells.AnalyzeMetacells.compute_metacells_3d_umap!

Metacells.AnalyzeMetacells.compute_vector_of_correlation_between_cells_and_punctuated_metacells_per_gene!

Metacells.AnalyzeMetacells.compute_vector_of_n_cells_per_metacell!

Metacells.AnalyzeMetacells.compute_vector_of_total_UMIs_per_metacell!

Metacells.AnalyzeMetacells.compute_vector_of_type_per_cell_by_metacells!

Metacells.AnalyzeMetacells.compute_vector_of_type_per_metacell_by_cells!