Analyze Modules

Metacells.AnalyzeModules — Module

Do simple gene module analysis.

Counts

Metacells.AnalyzeModules.compute_blocks_modules_n_genes! — Function

function compute_blocks_modules_n_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of genes in each gene module in each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Matrices

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ n_genes ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of genes in each gene module in each block.

Metacells.AnalyzeModules.compute_blocks_modules_n_covered! — Function

function compute_blocks_modules_n_covered!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of covered genes in each gene module in each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

Matrices

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ n_covered ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of covered genes in each gene module in each block.

Metacells.AnalyzeModules.compute_blocks_n_found_modules! — Function

function compute_blocks_n_found_modules!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of found gene modules for each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

Matrices

block, module @ is_found ::Bool (required): Whether each gene module is found in each block.

Outputs

Vectors

block @ n found modules ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of found modules for each block.

Metacells.AnalyzeModules.compute_blocks_n_used_modules! — Function

function compute_blocks_n_used_modules!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of used gene modules for each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

Matrices

block, module @ is_used ::Bool (required): Whether each gene module is used in the linear model of each block.

Outputs

Vectors

block @ n used modules ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of used modules for each block.

Masks

Metacells.AnalyzeModules.compute_blocks_modules_is_used! — Function

function compute_blocks_modules_is_used!(
    daf::DafWriter;
    min_used_module_covered_UMIs_fraction::Real = ```0.67```,
    overwrite::Bool = ```false```,
)::Nothing

Whether each gene module is used by each block. We only use modules that have a high fraction of covered gene UMIs (specifically, at least min_used_module_covered_UMIs_fraction ).

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

Matrices

block, module @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total UMIs of the covered genes in each gene module in each block.

block, module @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total UMIs of the genes in each gene module in each block.

Outputs

Matrices

block, module @ is_used ::Bool (guaranteed): Whether each gene module is used in the linear model of each block.

UMIs

Modules

Metacells.AnalyzeModules.compute_metacells_modules_total_UMIs! — Function

function compute_metacells_modules_total_UMIs!(daf::DafWriter; overwrite::Bool = false)::Nothing

The total UMIs of the genes of each module in each metacell.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

metacell, module @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the genes of each module in each metacell.

Metacells.AnalyzeModules.compute_metacells_modules_scaled_total_UMIs! — Function

function compute_metacells_modules_scaled_total_UMIs!(daf::DafWriter; overwrite::Bool = false)::Nothing

The total UMIs of the genes of each module in each metacell, scaled by divergence.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

metacell, module @ scaled total UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the genes of each module in each metacell, scaled by divergence.

Metacells.AnalyzeModules.compute_metacells_modules_covered_UMIs! — Function

function compute_metacells_modules_covered_UMIs!(daf::DafWriter; overwrite::Bool = false)::Nothing

The total UMIs of the covered genes of each module in each metacell.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

metacell, module @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the covered genes of each module in each metacell.

Metacells.AnalyzeModules.compute_metacells_modules_scaled_covered_UMIs! — Function

function compute_metacells_modules_scaled_covered_UMIs!(daf::DafWriter; overwrite::Bool = false)::Nothing

The total UMIs of the covered genes of each module in each metacell, scaled by divergence.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

metacell, module @ scaled covered UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the covered genes of each module in each metacell, scaled by divergence.

Blocks

Metacells.AnalyzeModules.compute_blocks_modules_total_UMIs! — Function

function compute_blocks_modules_total_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of the genes in each gene module in each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the genes in each gene module in each block.

Metacells.AnalyzeModules.compute_blocks_modules_covered_UMIs! — Function

function compute_blocks_modules_covered_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of the covered genes in each gene module in each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the covered genes in each gene module in each block.

Metacells.AnalyzeModules.compute_blocks_modules_scaled_total_UMIs! — Function

function compute_blocks_modules_scaled_total_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of the genes in each gene module in each block, scaled by divergence.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ scaled total UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the genes in each gene module in each bloc, scaled by divergence..

Metacells.AnalyzeModules.compute_blocks_modules_scaled_covered_UMIs! — Function

function compute_blocks_modules_scaled_covered_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of the covered genes in each gene module in each block, scaled by divergence.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

Outputs

Matrices

block, module @ scaled covered UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total UMIs of the covered genes in each gene module in each block, scaled by divergence.

Fractions

Modules

Linear Fractions

Metacells.AnalyzeModules.compute_metacells_modules_linear_fractions! — Function

function compute_metacells_modules_linear_fractions!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The estimated linear fraction of the UMIs of the genes of each module in each metacell, out of the total UMIs.

Metacells.AnalyzeModules.compute_metacells_modules_scaled_linear_fractions! — Function

function compute_metacells_modules_scaled_linear_fractions!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The estimated linear fraction of the UMIs of the genes of each module in each metacell, out of the total UMIs, scaled by divergence.

Metacells.AnalyzeModules.compute_metacells_modules_log_linear_fractions! — Function

function compute_metacells_modules_log_linear_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = false,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of the genes of each module in each metacell, out of the total UMIs. This adds the gene_fraction_regularization to deal with zero fractions.

Metacells.AnalyzeModules.compute_metacells_modules_log_scaled_linear_fractions! — Function

function compute_metacells_modules_log_scaled_linear_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = false,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of the genes of each module in each metacell, out of the total UMIs, scaled by divergence. This adds the gene_fraction_regularization to deal with zero fractions.

Linear Covered Fractions

Metacells.AnalyzeModules.compute_metacells_modules_linear_covered_fractions! — Function

function compute_metacells_modules_linear_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The estimated linear fraction of the UMIs of the covered genes of each module in each metacell, out of the total covered UMIs.

Metacells.AnalyzeModules.compute_metacells_modules_scaled_linear_covered_fractions! — Function

function compute_metacells_modules_scaled_linear_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The estimated linear fraction of the UMIs of the covered genes of each module in each metacell, out of the total covered UMIs, scaled by divergence.

Metacells.AnalyzeModules.compute_metacells_modules_log_linear_covered_fractions! — Function

function compute_metacells_modules_log_linear_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = false,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of the covered genes of each module in each metacell, out of the total covered UMIs. This adds the gene_fraction_regularization to deal with zero fractions.

Metacells.AnalyzeModules.compute_metacells_modules_log_scaled_linear_covered_fractions! — Function

function compute_metacells_modules_log_scaled_linear_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = false,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of the covered genes of each module in each metacell, out of the total covered UMIs, scaled by divergence. This adds the gene_fraction_regularization to deal with zero fractions.

Index

Metacells.AnalyzeModules

Metacells.AnalyzeModules.compute_blocks_modules_covered_UMIs!

Metacells.AnalyzeModules.compute_blocks_modules_is_used!

Metacells.AnalyzeModules.compute_blocks_modules_n_covered!

Metacells.AnalyzeModules.compute_blocks_modules_n_genes!

Metacells.AnalyzeModules.compute_blocks_modules_scaled_covered_UMIs!

Metacells.AnalyzeModules.compute_blocks_modules_scaled_total_UMIs!

Metacells.AnalyzeModules.compute_blocks_modules_total_UMIs!

Metacells.AnalyzeModules.compute_blocks_n_found_modules!

Metacells.AnalyzeModules.compute_blocks_n_used_modules!

Metacells.AnalyzeModules.compute_metacells_modules_covered_UMIs!

Metacells.AnalyzeModules.compute_metacells_modules_linear_covered_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_linear_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_log_linear_covered_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_log_linear_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_log_scaled_linear_covered_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_log_scaled_linear_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_scaled_covered_UMIs!

Metacells.AnalyzeModules.compute_metacells_modules_scaled_linear_covered_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_scaled_linear_fractions!

Metacells.AnalyzeModules.compute_metacells_modules_scaled_total_UMIs!

Metacells.AnalyzeModules.compute_metacells_modules_total_UMIs!