AnnData Format

Metacells.AnnDataFormat — Module

Import and export cells and metacells data from/to h5ad files. This allows moving data between the old Python/C++ based AnnData world and the brave new Julia based Daf world.

The expected flow is as follows:

Create a Daf repository for the raw input cells and import the raw cells h5ad into it using import_cells_h5ad! .

Alternatively, import just the clean cells h5ad into it, but that would be leaving out some of the data so is not recommended.

Create another Daf repository for the metacells, chain it to the cells repository, and import the metacells into it using import_metacells_h5ad! . Give this the cells-with-metacells data h5ad - only the assignment of cells to metacells will be imported from it.

If you have any per-cell or per-gene computed data in the cells-with-metacells data h5ad (unlikely, as computed data typically goes into the metacells h5ad ), import it into the chained (metacells) repository using import_cells_h5ad! .

Create a type axis in the chained metacells Daf repository using reconstruct_type_axis! .

Metacells.AnnDataFormat.import_cells_h5ad! — Function

function import_cells_h5ad!(
    daf::DafWriter;
    cells_h5ad::AbstractString,
    copy_data::Maybe{CopyAnnData} = ```nothing```,
    bestify::Bool = ```true```,
    min_sparse_saving_fraction::AbstractFloat = ```0.25```,
    overwrite::Bool = ```false```,
    insist::Bool = ```false```,
)::Nothing

Import an AnnData based cells dataset into a destination daf data set. Ideally you'd copy the full (raw) cells into an empty Daf repository. Then, you'd treat this repository as read-only, and copy the metacells data using import_metacells_h5ad! into a separate Daf repository chained with the read-only cells repository. This allows separate alternative metacells computations to share the read-only cells data.

You can copy an h5ad file containing just the clean cells on top of the raw cells data, to capture any data computed during or after computing the metacells. You may need to specify the copy_data to specify defaults for values of properties that exist only for clean cells and/or genes. Or, you can skip copying the raw data altogether, copying just the clean data into the base cells repository, though this is less recommended as you are needlessly discarding data that may prove to be useful later.

The bestify , min_sparse_saving_fraction , overwrite , and insist have their usual meaning from Daf 's copying functions.

When copying, we apply the following general rules:

A something_gene per-gene property and/or something_cell per-cell property are renamed to is_something , (and given a default of false ), because Daf (unlike AnnData ) has no problem with properties with the same name for different axes.

Similarly {gene,cell}[s]_something_module and something_{gene,cell}[s]_module properties are renamed to something_module . We add 1 to the value and store the results in a UInt32 ; that is, in Daf , module indices are 1-based, and 0 is "no module".

Any something_umis is renamed to something_UMIs , given a default of 0, and stored as a UInt32 .

And we make the following special exceptions:

Scalars:

We do not copy the __name__ scalar.

All other scalars are copied as-is.

Per-cell-per-gene:

The X matrix is renamed to UMIs , and stored as a UInt32 .

No other per-cell-per-gene matrix is copied by default.

Per-gene:

The per-gene correction_factor is given the default value 0.

The fitted per-gene vector is renamed to is_fitted and given the default false .

The significant_inner_folds_count per-gene property is stored as a UInt32 and given the default 0.

The full_gene_index property is not copied. Either you import the full (raw) data or you don't.

All other vectors are copied as-is.

Per-cell:

The full_cell_index property is not copied. Either you import the full (raw) data or you don't.

The metacell , metacell_name , metacell_level , most_similar and most_similar_name per-cell properties are not copied. To import these, use import_metacells_h5ad! .

All other vectors are copied as-is.

Note

It is common to manually call reconstruct_axis! on the result to create additional axes (e.g., if the cells were collected from a set of batches and some properties are actually per-batch).

Metacells.AnnDataFormat.import_metacells_h5ad! — Function

function import_metacells_h5ad!(
    daf::DafWriter;
    cells_h5ad::AbstractString,
    metacells_h5ad::AbstractString,
    copy_data::Maybe{CopyAnnData} = ```nothing```,
    bestify::Bool = ```true```,
    min_sparse_saving_fraction::AbstractFloat = ```0.25```,
    overwrite::Bool = ```false```,
    insist::Bool = ```false```,
)::Nothing

Import an AnnData based metacells dataset into a destination daf data set. It is expected that you have first imported the per-cell data. Ideally you'd create a new empty repository for the metacells data and chained it on top of the per-cell repository, which you'd keep read-only to allow sharing it when you (inevitably) compute different metacells for it.

This behaves similarly to import_cells_h5ad! , specifically the generic rules (except that we copy per-metacell properties and not per-cell properties so the rules are adjusted accordingly), and we make the following special exceptions:

Per-metacell-per-gene:

The X matrix is renamed to fraction and always stored as Float32 .

The corrected_fraction matrix is always stored as Float32 .

The essential matrix is renamed to is_essential .

The essential , fitted , and misfit matrices are renamed to is_essential , is_fitted and is_misfit , respectively.

The inner_fold , inner_stdev_log , projected_fold , projected_fraction matrices are always stored as Float32 .

The total_umis matrix is renamed to total_UMIs and always stored as UInt32 .

The zeros matrix is always stored as UInt32 .

All other matrices are copied as-is.

Scalars and Per-gene:

Same as in import_cells_h5ad!

Per-cell:

The only properties we copy per cell are metacell_name (renamed to metacell with a default of the empty string), and similarly most_similar_name (renamed to most_similar.metacell , same default). That's the only reason we have a cells_h5ad parameter. You should therefore pass here the cells-with-metacells and not the clean cells h5ad .

Per-metacell:

The metacells_level property is renamed to level .

The similar property is renamed to is_similar .

The type property is copied. If "the" type property of the metacells is different, use copy_data to rename it to type to match the Daf naming convention.

All other vectors are copied as-is.

Per-metacell-per-metacell:

The obs_outgoing_weights matrix is renamed to outgoing_weights and always stored as Float32 .

All other matrices are copied as-is.

Note

It is common to manually call reconstruct_type! on the result to create a type axis.

Metacells.AnnDataFormat.reconstruct_type_axis! — Function

reconstruct_type!(
    daf::DafWriter,
    base_axis::AbstractString = ```"metacell"```,
    type_property::AbstractString = ```"type"```,
    type_axis::AbstractString = ```"type"```,
    empty_type::Maybe{AbstractString} = ```nothing```,
    type_colors_csv::Maybe{AbstractString} = ```nothing```,
    implicit_properties::Maybe{AbstractSet{<:AbstractString}} = ```nothing```,
    skipped_properties::Maybe{AbstractSet{<:AbstractString}} = ```2 x Str (Set)```,
    properties_defaults::Maybe{Dict} = ```nothing```,
)::Nothing

Create a type axis after importing data containing type annotations. By default this will look for a type per metacell, but if you have type annotation per cell (which is not simply the type of the metacell they belong to), you can also use this for the cells.

By default this assumes that you have imported "the" type annotation to a property called "type", and that you would like the new type axis to be called "type" as well. If you want to import secondary types (or per-cell types), change these via the type_property and type_axis parameters.

If the type is equal to empty_type it is replaced with the empty string to match the Daf conventions for "no value" for string properties. Any set of per-gene properties named something_gene_of_type is converted to a per-gene-per-type matrix called is_something (with a default of false ).

Otherwise, this is mostly just a wrapper for reconstruct_axis! . It can be further enhanced by specifying a type_colors_csv file mapping type names to colors. This should be a comma or tab separated file containing at least two columns, one named "color" and one with the same name as the type_property . If this CSV file contains types that aren't actually used in the data, you will have to specify a default value for any other per-type property in properties_defaults .

Note

Most metacells data has type annotations and colors associated with types, so it is highly recommended you invoke this to capture these into the Daf repository. This will enable all types (:-) of downstream processing, coloring graphs, etc.

Metacells.AnnDataFormat.CopyAnnData — Type

Specify how to copy data from AnnData to Daf . The key is simply a vector or matrix name (ignoring axes), and the value is either nothing to ignore the data, or a tuple with the name of the destination Daf property and an optional value to use for missing entries (raw-only cells and/or genes).

Index

Metacells.AnnDataFormat

Metacells.AnnDataFormat.CopyAnnData

Metacells.AnnDataFormat.import_cells_h5ad!

Metacells.AnnDataFormat.import_metacells_h5ad!

Metacells.AnnDataFormat.reconstruct_type_axis!