Blocks

Metacells.Blocks — Module

Approximate the manifold of actual cell states (captured by metacells) using linear programs in each local region.

Metacells.Blocks.compute_covered_fractions! — Function

function compute_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```2.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

Compute normalized fractions for each metacell based only on the covered genes. By changing the denominator to only this smaller set of genes, we remove effects of possibly high-expression "irrelevant" gene programs (e.g., cell cycle). In general whenever we talk about gene "fractions" we have to keep in mind what is the total the fractions are out of, and ensure it is fit for purpose (and, of course, ensure the same divergence factors are used for all the data). Here we are going to be using these fractions for creating a distance measure in the manifold, which will be used for both deciding what is a "local" region and for estimating errors of reconsctructing the manifold, taking into consideration only the covered genes.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

gene @ divergence ::Union{Float32, Float64} (optional): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

Matrices

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

metacell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of the covered genes in each metacell.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (guaranteed): The relative fraction of covered genes.

gene, metacell @ scaled log covered_fraction ::Union{Float32, Float64} (guaranteed): The log of the relative fraction of covered genes, scaled by divergence.

Metacells.Blocks.compute_blocks! — Function

function compute_blocks!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```2.0e-5```,
    min_significant_gene_UMIs::Integer = ```40```,
    fold_confidence::AbstractFloat = ```0.9```,
    max_block_span::Real = ```2```,
    max_principal_components = ```40```,
    min_blocks_in_neighborhood::Integer = ```4```,
    min_metacells_in_neighborhood::Integer = ```20```,
    min_covered_UMIs_in_neighborhood::Integer = ```2000000```,
    cross_validation_parts::Integer = ```5```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Group the metacells into blocks where the metacells in each one are "similar" in all the skeleton genes. That is, each block is an approximation of a single cell state, assuming the skeleton genes adequately predict the rest of the genes. Then, group these blocks into overlapping small tight neighborhoods, and larger environments, such that each block's environment can be reasonably approximated using a local linear model, evaluated using cross-validation on the RMSE in the neighborhood at its core.

First, we compute metacell-metacell distances based on the maximal fold factor between is_skeleton genes. The fold factor is log (base 2) of the gene expression using the gene_fraction_regularization . For computing this fold factor, we ignore "insignificant" genes whose total UMIs in the compared metacells isn't at least min_significant_gene_UMIs . We also we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells, Two metacells can only belong to the same block if the final fold factor is at most max_block_span in all the (significant, skeleton) genes.

We then compute block-block distances which are the mean distance between the metacells of the blocks. Using these distances, we define a tight neighborhood of each block containing at least min_blocks_in_neighborhood , min_metacells_in_neighborhood and min_covered_UMIs_in_neighborhood (note that the latter only counts UMIs of skeleton genes).

Having computed the neighborhoods, we expand them (using the same block-block distances) as long as computing a local linear model for the environment gives a better approximation of the metacells in the neighborhood. This local model starts with computing max_principal_components . These are assumed to over-fit the solution, so we use cross-validation (in cross_validation_parts ) to pick a subset of these principal components, which hopefully approximates the true dimensionality of the local data.

Note

There is no attempt to reconcile the local model of neighboring blocks. In general the linear model computed here isn't very useful for interpreting the data as principal components for noisy high-dimensional data are pretty opaque.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

Matrices

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): The relative fraction of covered genes.

gene, metacell @ scaled log covered_fraction ::Union{Float32, Float64} (required): The log of the relative fraction of covered genes, scaled by divergence.

Outputs

Axes

block (guaranteed): Distinct groups of metacells with "very close" estimated cell state.

principal_component (guaranteed): A principal component of a local linear program.

Vectors

metacell @ block ::AbstractString (guaranteed): The unique block each metacell belongs to.

block @ n principal components ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of used principal components of the local linear model for each block.

block @ linear_RMSE ::Union{Float32, Float64} (guaranteed): The root mean squared error of predicting the covered genes using the linear model.

block @ linear_XRMSE ::Union{Float32, Float64} (guaranteed): The cross-validated root mean squared error of predicting the covered genes using the linear model.

Matrices

block, block @ is in neighborhood ::Bool (guaranteed): For each block, the mask of nearby blocks in its immediate neighborhood.

block, block @ is in environment ::Bool (guaranteed): For each block, the mask of nearby blocks in its linear environment.

block, principal component @ is used ::Bool (guaranteed): Whether each principal component is used by each block.

block, gene @ mean scaled log covered fraction ::Union{Float32, Float64} (guaranteed): The mean in the metacells of the block of the log of the relative fraction of covered genes, scaled by divergence.

Tensors

block; principal component, gene @ skeleton coefficient ::Union{Float32, Float64} (guaranteed): The coefficient of each skeleton gene for computing a principle component.

block; principal component, gene @ covered coefficient ::Union{Float32, Float64} (guaranteed): The coefficient of each principal component for computing each covered (non-skeleton) gene.

Metacells.Blocks.sharpen_metacells! — Function

TODOX

Index

Metacells.Blocks

Metacells.Blocks.compute_blocks!

Metacells.Blocks.compute_covered_fractions!

Metacells.Blocks.sharpen_metacells!