Boxes

Metacells.Boxes — Module

Given a set of raw metacells, partition them into boxes such that all metacells in the same box are within some (fold factor) radius of each other. The centroids of these boxes can serve as a representation of the cell state manifold which is less sensitive to oversampling of common cell states. Group these boxes in overlapping neighborhoods of "similar" boxes for further analysis.

Metacells.Boxes.compute_boxes! — Function

function compute_boxes!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    fold_confidence::AbstractFloat = 0.9,
    max_box_span::AbstractFloat = 2.0,
    max_neighborhood_span::AbstractFloat = 0.01,
    correlation_confidence::AbstractFloat = 0.9,
    max_deviant_genes_fraction::AbstractFloat = 0.01,
    overwrite::Bool = false,
)::Nothing

Partition raw metacells into distinct boxes, and boxes into overlapping neighborhoods.

Initial boxes and neighborhoods are computed in a first round, and then refined in a series of followup rounds.

In each round, we compute a distance between each two metacells. This is based on the fold factor between the expression level of each (relevant) gene in the metacells. The fold factor is the absolute value of the difference in the log (base 2) of the fraction of the gene in the metacells. This log is computed with the gene_fraction_regularization (by default, 1e-5 ). Since the fraction of the gene is a random variable, we decrease the high fraction and increase the low fraction by a factor based on the fold_confidence of the test (by default, 0.9), assuming a multinomial distribution. In addition, if the sum of the total UMIs of the gene in both metacells is less than min_significant_gene_UMIs (by default, 40 ), we ignore this fold factor as insignificant. Finally, genes, we reduce the fold factor using the gene's divergence factor. In the first round, we simply count the number of genes whose fold factor is more than max_box_span (for computing boxes) and max_box_span + max_neighborhood_span (for computing neighborhoods). In the followup rounds, we use the maximal gene fold, for genes that are correlated in the vicinity of the metacells (see below). Finally, in the followup rounds, we only compare a base metacell with other metacells which were in the same neighborhood in the previous round.

We use hierarchical clustering to partition the metacells to distinct boxes, such that the maximal distance between any metacells in the box is bounded. In the first round, this bound is the max_deviant_genes_fraction out of the total number of genes. In the followup rounds, this is the max_box_span .

For each box, we compute a main neighborhood of other boxes such that the maximal distance between any metacells in the neighborhood is bounded. In the first round, this bound is again the maximal number of deviant genes (this time, using the increased fold distance computed above). In the followup rounds, this is the max_box_span plus the max_neighborhood_span . These neighborhoods may overlap. The main neighborhoods of different boxes may even be identical.

For each box, we compute the set of genes which are correlated (using the correlation_confidence ) with some other gene(s) in its main neighborhood.

If the new sets of correlated genes only differ up to max_convergence_fraction , consider this to have converged. Otherwise, repeat from step 2.

If overwrite is set, the results will replace any previously computed boxes and neighborhoods.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Outputs

Axes

box (guaranteed): Distinct groups of metacells with "very close" estimated cell state.

neighborhood (guaranteed): Overlapping groups of boxes with "close" estimated cell states.

Vectors

metacell @ box ::AbstractString (guaranteed): The unique box each metacell belongs to.

box @ neighborhood.main ::AbstractString (guaranteed): The unique main neighborhood of each box.

neighborhood @ span ::Union{Float32, Float64} (guaranteed): The span (fold factor) used to compute the neighborhood.

Matrices

gene, neighborhood @ is_correlated ::Bool (guaranteed): Which genes are correlated in each neighborhood.

box, box @ distance ::Union{Float32, Float64} (guaranteed): The distance (fold factor) between the most different metacell genes between the boxes.

box, neighborhood @ is_member ::Bool (guaranteed): A mask of the member boxes of each neighborhood.

Metacells.Boxes.compute_boxes_data! — Function

function compute_boxes_data!(daf::DafWriter)::Nothing

Compute aggregated per-box data based on per-metacell data. This is separated from compute_boxes! to allow computing this data for all genes (instead of just for the subset of genes fed to compute_boxes! ).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

box (required): Distinct groups of metacells with "very close" estimated cell state.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Outputs

Matrices

gene, box @ fraction ::Union{Float32, Float64} (guaranteed): The estimated fraction of the UMIs of each gene in each box.

gene, box @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs used to estimate the fraction of each gene in each box.

Index

Metacells.Boxes

Metacells.Boxes.compute_boxes!

Metacells.Boxes.compute_boxes_data!