Collect Metacells

Metacells.Collect — Module

Collect metacells.

Metacells.Collect.collect_metacells! — Function

collect_metacells(
    daf::DafWriter;
    UMIs_regularization::AbstractFloat = ```0.0625```,
    min_downsamples::Integer = ```750```,
    min_downsamples_quantile::AbstractFloat = ```0.05```,
    max_downsamples_quantile::AbstractFloat = ```0.5```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Given an assignment of cells to metacell, compute an estimation of the fraction of UMIs of each gene for each metacell.

The naive way to do this would be to just take the total UMIs of the gene out of the total UMIs of the metacell. However, this method has a weakness; a single strong cell with a "very different" fraction of the gene will dominate (that is, the method is sensitive to outliers).

Instead, we take the geometric mean of the fractions of the gene in the cells. This however raises a few issues we need to deal with:

Genes with zero UMIs are a problem; to combat this, we add a UMIs_regularization factor when computing the fractions, take the geomean, and subtract the regularization at the end (so all-zero genes will still get a zero overall fraction).

We want to give more weight to cells with more UMIs, so we use a scaled geomean. The weight we give to each cell is the log of the total number of UMIs in it.

The geomean fractions of all the genes in a metacell do not sum to one, so we scale them. This has the unfortunate side effect that the end result does not obey a nice relation to the linear fraction of the gene in each of the cells; in particular, the final result might be higher than all of these per-cell fractions (ouch).

This raises an important point about the whole "fraction of UMIs of a gene in a cell" concept, which is that it is highly dependent on the set of genes you pick to compute the fraction out of (that is, the denominator). This places restrictions on how you should use these fractions:

You can't just compare these fractions between two arbitrary data sets, as the denominators aren't the same. You must identify the set of common genes, and renormalize the fractions to sum to one in this subset, in both data sets. Only then can you meaningfully compare the results.

Even if you compare two data sets with the same set of genes (or even two subsets of the same data set, such as two "cell types"), if a specific "gene program" has a very high total expression in only one of them, then all other genes will artificially appear to be lower.

This is why we recommend excluding ribosomal genes from the data sets; they can take up anything between almost none to over two thirds of the total UMIs, and this varies between "cell types". This means that if they are included in the denominator, the fractions of otherwise "identical" genes will appear to differ by a factor of up to 3X

Luckily, most gene programs we deal with have a total expression of a few percent at most, so this effect is negligible. However if you identify a gene program with a total expression higher than, say, 10%, you should consider its effect on the rest of the genes.

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the genes in each cell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

Matrices

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs used to estimate the fraction of each gene in each metacell.

gene, metacell @ fraction ::Union{Float32, Float64} (guaranteed): The estimated fraction of the UMIs of each gene in each metacell.

Index

Metacells.Collect

Metacells.Collect.collect_metacells!