Compute Approximation

Metacells.ComputeApproximation — Module

Compute a local linear approximation to the manifold based on gene modules.

Metacells.ComputeApproximation.compute_blocks_modules_base_covered_fractions! — Function

function compute_blocks_modules_base_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The base fractions used by the gene modules analysis. This is based on the per-gene base_covered_fraction .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

Matrices

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

block, gene @ base covered fraction ::Union{Float32, Float64} (required): The base covered fractions used by the principal component analysis..

Outputs

Matrices

block, module @ base covered fraction ::Union{Float32, Float64} (guaranteed): The base covered fractions used by the gene modules analysis..

Metacells.ComputeApproximation.compute_approximation! — Function

function compute_approximation!(
    daf::DafWriter;
    cross_validation_parts::Integer = ```5```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

For each block, compute a local linear model for its environment, based on the used gene modules. This is a simple least-squares approximation predicting the fraction of the covered genes based on the fraction of the gene modules (using the base_covered_fraction of the modules in the environment). This is assumed to over-fit the solution, so we use cross-validation (in cross_validation_parts ) using subsets of the metacells to compute both the RMSE and XRMSE to estimate this.

Note

This uses the virtual metacell_gene_covered_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

module (required): A local (per block) gene module.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): Some estimated fraction of the UMIs of each covered gene in each metacell.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, block @ is in neighborhood ::Bool (required): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (required): The index of the gene module each gene belongs to in the environment of each block.

block, gene @ base covered fraction ::Union{Float32, Float64} (required): The base covered fractions used by the principal component analysis..

block, module @ base covered fraction ::Union{Float32, Float64} (required): The base covered fractions used by the gene modules analysis..

block, module @ is_used ::Bool (required): Whether each gene module is used in the linear model of each block.

Outputs

Vectors

block @ modules_RMSE ::Union{Float32, Float64} (guaranteed): The root mean squared error of predicting the covered genes using the gene modules in the vicinity of the block.

block @ modules_XRMSE ::Union{Float32, Float64} (guaranteed): The cross-validated root mean squared error of predicting the covered genes using the gene modules in the vicinity of the block.

Tensors

block; module, gene @ covered_coefficient ::Union{Float32, Float64} (guaranteed): The coefficient of each gene module for computing each covered gene.

Index

Metacells.ComputeApproximation

Metacells.ComputeApproximation.compute_approximation!

Metacells.ComputeApproximation.compute_blocks_modules_base_covered_fractions!

Model