Compute Blocks

Metacells.ComputeBlocks — Module

Group "very close" metacells into disjoint blocks.

Metacells.ComputeBlocks.compute_metacells_blocks! — Function

function compute_metacells_blocks!(
    daf::DafWriter;
    num_blocks::Integer = ```nothing```,
    max_block_span::Real = ```2```,
    prefix::AbstractString = ```"B"```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set the block_axis and vector_of_block_per_metacell . We group the metacells into blocks where the metacells in each one are "similar" in all the skeleton genes. That is, each block is an approximation of a single cell state, assuming the skeleton genes adequately predict the rest of the genes.

If num_blocks is not specified, we decide on the number of blocks by using hierarchical clustering based on the maximal skeleton fold distance between the metacells in each cluster of max_block_span (that is, hclust with :complete linkage using the matrix_of_max_skeleton_fold_distance_between_metacells .

The actual blocks are computed using hierarchical clustering with :ward linkage based on the euclidean skeleton distance between the metacells (that is, using the matrix_of_euclidean_skeleton_fold_distance_between_metacells ).

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

metacell, metacell @ max skeleton fold_distance ::Union{Float32, Float64} (required): The maximal significant fold factor between the fraction of skeleton genes between the metacells.

metacell, metacell @ euclidean skeleton fold_distance ::Union{Float32, Float64} (required): The Euclidean distance between the log of the fraction of the skeleton genes between the metacells.

Outputs

Axes

block (created): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (created): The unique block each metacell belongs to.

Index

Metacells.ComputeBlocks

Metacells.ComputeBlocks.compute_metacells_blocks!