Compute Modules

Metacells.ComputeModules — Module

Group "very correlated" genes modules in each local environment.

Metacells.ComputeModules.compute_blocks_modules! — Function

function compute_blocks_modules!(
    daf::DafWriter;
    min_module_downsampled_UMIs::Integer = ```8```,
    min_module_strong_cells::Integer = ```12```,
    min_merge_fragments_correlation::Real = ```0.5```,
    min_merge_modules_correlation::Real = ```0.8```,
    min_downsamples::Integer = ```750```,
    min_downsamples_quantile::AbstractFloat = ```0.05```,
    max_downsamples_quantile::AbstractFloat = ```0.5```,
    overwrite::Bool = ```false```,
)::Nothing

Group the genes into modules based on their correlation.

For each block:

Compute the correlation of the expression of the genes across all metacells in the block's environment.

Cluster the genes based on this correlation, using the :complete distance measure.

Bottom-up group genes into fragments and then into complete modules. Fragments have at least min_merge_fragments_correlation between all their genes. At some point, the fragment will contain enough genes so there will be at least min_module_strong_cells . A "strong" cell is expected to have at least min_module_downsampled_UMIs , if we downsample the cells in the environment using min_downsamples , min_downsamples_quantile , and max_downsamples_quantile . If the divergence is specified per gene, then it is applied to the UMIs. When this happens, we mark the fragment (which might be a single gene!) as a valid module. We will only merge two such modules if the correlation between all their genes is at least min_merge_modules_correlation .

Having merged some of the genes into modules, we repeat steps 2 and 3 for the unmerged genes until no new modules are found.

Most of the genes will not end up in a valid module, and are given the module index 0. The other genes are given a 1-based module index. We need to store a matrix of module indices instead of module names due to Daf limitations (no matrices of strings).

Note

This uses the virtual metacell_gene_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...). In addition, you might want to restrict the set of genes that are candidates to be included in the module, for example just to is_covered genes.

Inputs

Axes

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

gene @ divergence ::Union{Float32, Float64} (optional): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

gene, metacell @ fraction ::Union{Float32, Float64} (required): Some estimated fraction of the UMIs of each gene in each metacell.

block, gene @ is environment marker ::Bool (required): A mask of genes that distinguish between cell states in each environment.

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

Outputs

Axes

module (guaranteed): A local (per block) gene module.

Matrices

block, gene @ module_index ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The index of the gene module each gene belongs to in the environment of each block.

block, module @ is_found ::Bool (guaranteed): Whether each gene module is found in each block.

block, module @ min gene correlation ::Union{Float32, Float64} (guaranteed): The minimal correlation between the genes of each gene module of each block.

Index