Defaults

Metacells.Defaults — Module

Default values for parameters.

Metacells.Defaults.GENE_FRACTION_REGULARIZATION — Constant

When computing log (base 2) of the fraction of the gene expression out of the total, we use this regularization factor by default top avoid zero values.

Metacells.Defaults.MIN_SIGNIFICANT_GENE_UMIS — Constant

When comparing gene expression levels (e.g., when computing fold factors), we do not consider the result to be significant unless at least this number of UMIs were used to compute the compared value(s).

Index

Metacells.Defaults

Metacells.Defaults.GENE_FRACTION_REGULARIZATION

Metacells.Defaults.MIN_SIGNIFICANT_GENE_UMIS