Identify Genes

Metacells.IdentifyGenes — Module

Identify special genes.

Metacells.IdentifyGenes.identify_marker_genes! — Function

function identify_marker_genes!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1e-5,
    min_gene_range_fold::AbstractFloat = 2.0,
    noisy_gene_fold::AbstractFloat = 1.0,
    min_max_marker_gene_fraction::AbstractFloat = 1e-4,
    overwrite::Bool = false,
)::Nothing

Identify the genes that distinguish at least one metacell from the rest. Such genes are called "marker" genes as they (potentially) mark specific cell states. If overwrite , will overwrite an existing is_marker mask.

Compute the minimal and maximal expression level of each gene.

Select the genes whose fold factor (log2 of maximal over minimal value, using the gene_fraction_regularization is at least min_marker_gene_range_fold . For is_noisy genes, we require an additional noisy_gene_fold .

Identify the genes whose maximal expression is at least min_max_marker_gene_fraction .

CONTRACT

Metacells.IdentifyGenes.identify_lonely_genes! — Function

function identify_lonely_genes!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1e-5,
    max_lonely_gene_correlation::AbstractFloat = 0.5,
    max_lonely_gene_correlated::Integer = 2,
    overwrite::Bool = false,
)::Nothing

Identify genes that are uncorrelated with other genes. Such "lonely" genes make it difficult to identify groups of genes that act together. If overwrite , will overwrite an existing is_lonely mask.

Compute the log base 2 of the genes expression in each metacell (using the gene_fraction_regularization ).

Correlate this between all the pairs of genes.

Find the maximal correlation for each gene.

Identify the genes which have up to max_lonely_gene_correlated genes with a correlation of more than max_lonely_gene_correlation .

CONTRACT

Index

Metacells.IdentifyGenes

Metacells.IdentifyGenes.identify_lonely_genes!

Metacells.IdentifyGenes.identify_marker_genes!