Programs

Metacells.Programs — Module

Given a set of raw metacells, partition them into blocks such that all metacells in the same block are within some (fold factor) radius of each other. The centroids of these blocks can serve as a representation of the cell state manifold which is less sensitive to oversampling of common cell states. Group these blocks in overlapping neighborhoods of "similar" blocks for further analysis.

Metacells.Programs.compute_global_predictive_factors! — Function

function compute_global_predictive_factors!(  # untested
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1.0e-5
    max_principal_components::Integer = 30,
    rng::AbstractRNG = default_rng(),
)::Nothing

TODOY

Metacells.Programs.compute_local_predictive_factors! — Function

function compute_local_predictive_factors!(  # untested
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1.0e-5,
    max_principal_components::Integer = 40,
    fold_confidence::AbstractFloat = 0.9,
    max_block_span::AbstractFloat = 2.0,
    min_blocks_in_neighborhood::Integer = 4,
    min_metacells_in_neighborhood::Integer = 100,
    rng::AbstractRNG = default_rng(),
)::Nothing

Index

Metacells.Programs

Metacells.Programs.compute_global_predictive_factors!

Metacells.Programs.compute_local_predictive_factors!