Project Cells

Metacells.ProjectCells — Module

Project cells on an atlas with modules.

Metacells.ProjectCells.compute_cells_projection! — Function

compute_cells_projection(;
    query_daf::DafWriter,
    atlas_daf::DafReader,
    gene_fraction_regularization::Real = ```0.0001```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set [ vector_of_projected_block_per_cell ] and vector_of_projected_modules_z_score_per_cell . To pick the best metacell in the atlas_daf for each cell of the query_daf , we:

Pick a provisional block for each cell. This is the block with the minimal Euclidean distance of the non-lateral marker genes.

Then, for each cell, consider the expression level of the found modules of the provisional block's neighborhood. We use this to pick a metacell in the neighborhood which has the closest Euclidean distance.

This metacell might belong to a different block. If so, we repeat the process using that block's neighborhood modules.

If the resulting best match metacell is in the same (new) block, we accept it.

Otherwise, we just look for the closest metacell in the original block (using that block's modules) and settle for that.

Distances are computed on the log (base 2) of the gene expression using the gene_fraction_regularization to handle zero fractions. We also compute the z-score (final distance between the cell and projected metacell, minus the mean distance of the cells in the metacell, divided by their standard deviation). If this is too large, then the query cell isn't a good match for anything in the atlas.

Query

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

Vectors

cell @ is_excluded ::Bool (required): A mask of cells that are totally excluded from the analysis.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

cell @ projected_metacell ::AbstractString (created): The projected atlas cell for each metacell.

cell @ projected metacell modules z score ::Union{Float32, Float64} (created): The z score of the distance between each query cell and its projected atlas metacell modules.

cell @ projected_block ::AbstractString (created): The atlas block of the projected metacell for each query cell.

Atlas

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

gene (required): Sequenced genes.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ mean euclidean modules cells distance ::Union{Float32, Float64} (required): In each metacell, the mean of the euclidean distances of between its cells and the metacell's linear fraction of the modules of the metacell's block's neighborhood.

metacell @ std euclidean modules cells distance ::Union{Float32, Float64} (required): In each metacell, the standard deviation of the euclidean distances of between its cells and the metacell's linear fraction of the modules of the metacell's block's neighborhood.

Matrices

block, module @ is_found ::Bool (required): A mask of the modules that were found for each block.

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

block, module @ mean linear fraction in neighborhood_cells ::Union{Float32, Float64} (required): The mean of the linear fraction of each module in the cells of the neighborhood of each block.

block, module @ std linear fraction in neighborhood_cells ::Union{Float32, Float64} (required): The standard deviation of the linear fraction of each module in the cells of the neighborhood of each block.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, block @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Tensors

block; metacell, module @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each block's found module genes out of the total UMIs in each metacell.

Metacells.ProjectCells.compute_vector_of_correlation_between_cells_and_projected_metacells! — Function

function compute_vector_of_correlation_between_cells_and_projected_metacells!(;
    atlas_daf::DafReader,
    query_daf::DafWriter,
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set (in query) vector_of_correlation_between_cells_and_projected_metacells_per_gene .

Query

Inputs

Axes

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

Vectors

cell @ is_excluded ::Bool (required): A mask of cells that are totally excluded from the analysis.

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

cell @ projected_metacell ::AbstractString (required): The projected atlas cell for each metacell.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

gene @ correlation between cells and projected_metacells ::Union{Float32, Float64} (created): The correlation between query cells and their projected atlas metacells gene expression levels.

Atlas

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Metacells.ProjectCells.compute_matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block! — Function

compute_matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block!(
    query_daf::DafWriter,
    atlas_daf::DafReader,
    gene_fraction_regularization::AbstractFloat = ```0.0001```,
    min_neighborhood_query_cells::Integer = ```30```,
    overwrite::Bool = false,
)::Nothing

Compute and set matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block . If there are less than min_neighborhood_query_cells , we set this to zero. This will create projected_block_axis in the query_daf if necessary.

Query

Inputs

Axes

gene (required): Sequenced genes.

cell (required): Sequenced single cells.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

cell @ projected_metacell ::AbstractString (required): The projected atlas cell for each metacell.

cell @ projected_block ::AbstractString (required): The atlas block of the projected metacell for each query cell.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Matrices

gene, projected block @ correlation between neighborhood cells and projected_metacells ::Union{Float32, Float64} (created): The correlation between query cells and their projected atlas metacells of each gene's expression levels in each block's neighborhood.

Atlas

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

Matrices

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

gene, block @ is neighborhood marker ::Bool (optional): A mask of genes that distinguish between cell states in each block's neighborhood.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Index

Metacells.ProjectCells

Metacells.ProjectCells.compute_cells_projection!

Metacells.ProjectCells.compute_matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block!

Metacells.ProjectCells.compute_vector_of_correlation_between_cells_and_projected_metacells!