Spheres

Metacells.Spheres — Module

Given a set of raw metacells, partition them into spheres such that all metacells in the same sphere are within some (fold factor) radius of each other. The centroids of these spheres can serve as a representation of the cell state manifold which is less sensitive to oversampling of common cell states. Group these spheres in overlapping neighborhoods of "similar" spheres for further analysis.

Metacells.Spheres.compute_improved_spheres! — Function

function compute_improved_spheres!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9,
    max_sphere_diameter::AbstractFloat = 2.0,
    max_neighborhood_diameter::AbstractFloat = 2.0,
    noisy_gene_fold::AbstractFloat = 1.0,
    min_gene_correlation::AbstractFloat = 0.5,
)::Nothing

Partition raw metacells into distinct spheres, and spheres into overlapping neighborhoods, such that each sphere is homogeneous in the genes that participate in multi-gene behaviors in its main neighborhood.

This calls compute_spheres! and then `improve_spheres! . In the 1st call to compute_spheres! call, we allow up to max_deviant_genes_fraction of the genes to be deviant; in improve_spheres! , we forbid all genes from being deviant, as we expect to ignore all "irrelevant" genes as non-markers or lonely.

CONTRACT

Metacells.Spheres.compute_spheres! — Function

function compute_spheres!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9,
    max_sphere_diameter::AbstractFloat = 2.0,
    max_neighborhood_diameter::AbstractFloat = 2.0,
    noisy_gene_fold::AbstractFloat = 1.0,
    max_deviant_genes::Integer = 0,
    overwrite::Bool = false,
)::Nothing

Partition raw metacells into distinct spheres, and spheres into overlapping neighborhoods.

Note

This is a lower-level function; you probably want to call compute_improved_spheres! instead.

Compute a distance between each two metacells, defined as the maximal fold factor of any of the genes. This fold factor is the absolute value of the difference in the log base 2 of the fraction of the gene in both metacells, using the gene_fraction_regularization (by default, 1e-5 ). Since the fraction of the gene is a random variable, we decrease the high fraction and increase the low fraction by a factor based on the confidence of the test (by default, 0.9), assuming a multinomial distribution. In addition, if the sum of the total UMIs of the gene in both metacells is less than min_significant_gene_UMIs (by default, 40 ), we ignore this fold factor as insignificant. We also ignore the fold factors for genes that aren't is_marker in the vicinity of at least one of the metacells, or that are is_lonely in the vicinity of both. For is_noisy_gene_fold genes, we reduce the distance by noisy_gene_fold to allow for their bursty nature.

If max_deviant_genes is not zero, we redefine the distance between metacells to be the number of genes with a maximal fold distance above the max_sphere_diameter (for computing spheres) and max_neighborhood_diameter (for computing neighborhoods). Otherwise, we use the raw maximal fold distance.

Compute hierarchical clustering of the metacells using these distances and the complete linkage; that is, each cluster can be seen as a sphere of some diameter where all the metacells are within that sphere.

Use this hierarchical clustering to partition the metacells into as-large as possible spheres. If max_deviant_genes is zero (the default), then this is cutree using the max_sphere_diameter (by default, 2.0 , or 4x); otherwise, it is a cutree using the max_deviant_genes .

Compute for each sphere the set of other spheres in its main neighborhood. If max_deviant_genes is zero, this means the maximal distance between any pair of metacells is at most an additional max_neighborhood_diameter (by default, 2.0 , for a total neighborhood diameter of 16x). Otherwise, it means the maximal number of genes whose distance is more than this threshold in the metacells pair is at most max_deviant_genes . Neighborhoods can overlap; also, if the main neighborhoods defined by two spheres end up being identical, we unify them, so we may end up with less neighborhoods than spheres.

The distances to the other spheres in the main neighborhood of each sphere form a natural graph representation of the structure of the manifold of the distinct biological cell states captured by the metacells.

If overwrite is set, the results will replace any previously computed spheres and neighborhoods.

CONTRACT

Metacells.Spheres.improve_spheres! — Function

function improve_spheres!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9,
    max_sphere_diameter::AbstractFloat = 2.0,
    max_neighborhood_diameter::AbstractFloat = 2.0,
    noisy_gene_fold::AbstractFloat = 1.0,
    min_gene_correlation::AbstractFloat = 0.5,
)::Nothing

Improve the partition raw metacells into distinct spheres, and spheres into overlapping neighborhoods. The improved partition ensures no gene is not "too different" between the metacells in each sphere and spheres in each neighborhood; however, this only applies to genes that have at least min_gene_correlation with at least one other gene in the neighborhood.

Invoke identify_marker_genes! and identify_lonely_genes! for each neighborhood and store the results in is_marker and is_lonely of the genes for the metacells in sphere for which this is the main neighborhood.

Recompute the spheres with max_deviant_genes = 0 and overwrite = true . When computing the distance between each two metacells, this will only consider genes that are markers in either metacell and are not lonely in both. The result should be "improved" spheres as they will not be affected by genes that aren't part of some multi-gene behavior relevant to each sphere's neighborhood.

CONTRACT

Index

Metacells.Spheres

Metacells.Spheres.compute_improved_spheres!

Metacells.Spheres.compute_spheres!

Metacells.Spheres.improve_spheres!