Metacells

Metacells.Metacells — Module

The Metacells.jl package provides computational services for the metacells package, using Daf to hold the data. In the future, we'll ideally migrate all of the metacells package computations to this package, converting the Python package to a thin wrapper, and provide a similar thin R wrapper to provide metacell analysis from R as well. For now, Metacells.jl only provides a subset of the features of the Python metacells package, which requires users to convert data from AnnData (for the old features) to Daf (to the new features).

Index

Metacells.AnalyzeBlocks

Metacells.AnalyzeCells

Metacells.AnalyzeGenes

Metacells.AnalyzeMetacells

Metacells.AnalyzeModules

Metacells.AnnDataFormat

Metacells.ComputeBlocks

Metacells.ComputeModules

Metacells.Contracts

Metacells.Defaults

Metacells.Gmara

Metacells.Metacells

Metacells.ProjectCells

Metacells.SharpenMetacells

Metacells.Defaults.GENE_FRACTION_REGULARIZATION_FOR_CELLS

Metacells.Defaults.GENE_FRACTION_REGULARIZATION_FOR_METACELLS

Metacells.Defaults.MIN_SIGNIFICANT_GENE_UMIS

Metacells.Gmara.CACHE_DIR

Metacells.Gmara.TIMEOUT

Metacells.AnnDataFormat.CopyAnnData

Metacells.AnalyzeBlocks.compute_matrix_of_UMIs_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_correlation_with_most_between_base_neighborhood_cells_and_metacells_per_gene_per_base_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_in_neighborhood_per_block_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_distinct_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_is_neighborhood_marker_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_linear_fraction_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_log_linear_fraction_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_matrix_of_mean_euclidean_skeleton_fold_distance_between_blocks!

Metacells.AnalyzeBlocks.compute_matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_block_closest_by_pertinent_markers_per_cell!

Metacells.AnalyzeBlocks.compute_vector_of_n_cells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_metacells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_blocks_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_cells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_n_neighborhood_metacells_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_total_UMIs_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_total_neighborhood_UMIs_per_block!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_cells!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_block_by_metacells!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_cell_by_blocks!

Metacells.AnalyzeBlocks.compute_vector_of_type_per_metacell_by_blocks!

Metacells.AnalyzeCells.compute_vector_of_total_UMIs_per_cell!

Metacells.AnalyzeGenes.compute_vector_of_is_correlated_with_skeleton_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_is_marker_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_is_skeleton_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_marker_rank_per_gene!

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_regulator_per_gene!

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_transcription_factor_per_gene!

Metacells.AnalyzeGenes.rank_variables

Metacells.AnalyzeMetacells.compute_matrix_of_UMIs_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_euclidean_skeleton_fold_distance_between_metacells!

Metacells.AnalyzeMetacells.compute_matrix_of_linear_fraction_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_log_linear_fraction_per_gene_per_metacell!

Metacells.AnalyzeMetacells.compute_matrix_of_max_skeleton_fold_distance_between_metacells!

Metacells.AnalyzeMetacells.compute_vector_of_correlation_between_cells_and_punctuated_metacells_per_gene!

Metacells.AnalyzeMetacells.compute_vector_of_n_cells_per_metacell!

Metacells.AnalyzeMetacells.compute_vector_of_total_UMIs_per_metacell!

Metacells.AnalyzeMetacells.compute_vector_of_type_per_cell_by_metacells!

Metacells.AnalyzeMetacells.compute_vector_of_type_per_metacell_by_cells!

Metacells.AnalyzeModules.compute_matrix_of_n_genes_per_module_per_block!

Metacells.AnalyzeModules.compute_stats_of_euclidean_modules_cells_distance_per_metacell!

Metacells.AnalyzeModules.compute_stats_of_linear_fraction_in_neighborhood_cells_per_module_per_block!

Metacells.AnalyzeModules.compute_tensor_of_linear_fraction_per_block_per_module_per_metacell!

Metacells.AnalyzeModules.compute_vector_of_n_modules_per_block!

Metacells.AnnDataFormat.import_cells_h5ad!

Metacells.AnnDataFormat.import_metacells_h5ad!

Metacells.AnnDataFormat.reconstruct_type_axis!

Metacells.ComputeBlocks.compute_metacells_blocks!

Metacells.ComputeModules.compute_blocks_modules!

Metacells.Contracts.base_block_axis

Metacells.Contracts.block_axis

Metacells.Contracts.cell_axis

Metacells.Contracts.gene_axis

Metacells.Contracts.matrix_of_UMIs_per_gene_per_block

Metacells.Contracts.matrix_of_UMIs_per_gene_per_cell

Metacells.Contracts.matrix_of_UMIs_per_gene_per_metacell

Metacells.Contracts.matrix_of_confusion_by_closest_by_pertinent_markers_per_block_per_block

Metacells.Contracts.matrix_of_correlation_between_base_neighborhood_cells_and_punctuated_metacells_per_gene_per_base_block

Metacells.Contracts.matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block

Metacells.Contracts.matrix_of_correlation_between_neighborhood_cells_and_punctuated_metacells_per_gene_per_block

Metacells.Contracts.matrix_of_correlation_with_most_between_base_neighborhood_cells_and_metacells_per_gene_per_base_block

Metacells.Contracts.matrix_of_euclidean_skeleton_fold_distance_between_metacells

Metacells.Contracts.matrix_of_is_correlated_with_skeleton_in_neighborhood_per_gene_per_block

Metacells.Contracts.matrix_of_is_found_per_module_per_block

Metacells.Contracts.matrix_of_is_in_neighborhood_per_block_per_block

Metacells.Contracts.matrix_of_is_neighborhood_distinct_per_gene_per_block

Metacells.Contracts.matrix_of_is_neighborhood_marker_per_gene_per_block

Metacells.Contracts.matrix_of_linear_fraction_per_gene_per_block

Metacells.Contracts.matrix_of_linear_fraction_per_gene_per_metacell

Metacells.Contracts.matrix_of_log_linear_fraction_per_gene_per_block

Metacells.Contracts.matrix_of_log_linear_fraction_per_gene_per_metacell

Metacells.Contracts.matrix_of_max_skeleton_fold_distance_between_metacells

Metacells.Contracts.matrix_of_mean_euclidean_skeleton_fold_distance_between_blocks

Metacells.Contracts.matrix_of_mean_linear_fraction_in_neighborhood_cells_per_module_per_block

Metacells.Contracts.matrix_of_module_per_gene_per_block

Metacells.Contracts.matrix_of_module_status_per_gene_per_block

Metacells.Contracts.matrix_of_most_correlated_gene_in_neighborhood_per_gene_per_block

Metacells.Contracts.matrix_of_n_genes_per_module_per_block

Metacells.Contracts.matrix_of_std_linear_fraction_in_neighborhood_cells_per_module_per_block

Metacells.Contracts.metacell_axis

Metacells.Contracts.module_axis

Metacells.Contracts.projected_block_axis

Metacells.Contracts.tensor_of_linear_fraction_per_block_per_module_per_metacell

Metacells.Contracts.type_axis

Metacells.Contracts.vector_of_anchor_per_module

Metacells.Contracts.vector_of_base_block_per_metacell

Metacells.Contracts.vector_of_block_closest_by_pertinent_markers_per_cell

Metacells.Contracts.vector_of_block_per_metacell

Metacells.Contracts.vector_of_color_per_type

Metacells.Contracts.vector_of_correlation_between_cells_and_projected_metacells_per_gene

Metacells.Contracts.vector_of_correlation_between_cells_and_punctuated_metacells_per_gene

Metacells.Contracts.vector_of_excluded_UMIs_per_cell

Metacells.Contracts.vector_of_is_correlated_with_skeleton_per_gene

Metacells.Contracts.vector_of_is_excluded_per_cell

Metacells.Contracts.vector_of_is_excluded_per_gene

Metacells.Contracts.vector_of_is_forbidden_per_gene

Metacells.Contracts.vector_of_is_lateral_per_gene

Metacells.Contracts.vector_of_is_marker_per_gene

Metacells.Contracts.vector_of_is_mitochondrial_per_gene

Metacells.Contracts.vector_of_is_regulator_per_gene

Metacells.Contracts.vector_of_is_ribosomal_per_gene

Metacells.Contracts.vector_of_is_skeleton_per_gene

Metacells.Contracts.vector_of_is_transcription_factor_per_gene

Metacells.Contracts.vector_of_marker_rank_per_gene

Metacells.Contracts.vector_of_mean_euclidean_modules_cells_distance_per_metacell

Metacells.Contracts.vector_of_metacell_per_cell

Metacells.Contracts.vector_of_mitochondrial_UMIs_per_cell

Metacells.Contracts.vector_of_n_cells_per_block

Metacells.Contracts.vector_of_n_cells_per_metacell

Metacells.Contracts.vector_of_n_metacells_per_block

Metacells.Contracts.vector_of_n_modules_per_block

Metacells.Contracts.vector_of_n_neighborhood_blocks_per_block

Metacells.Contracts.vector_of_n_neighborhood_cells_per_block

Metacells.Contracts.vector_of_n_neighborhood_metacells_per_block

Metacells.Contracts.vector_of_projected_block_per_cell

Metacells.Contracts.vector_of_projected_metacell_per_cell

Metacells.Contracts.vector_of_projected_modules_z_score_per_cell

Metacells.Contracts.vector_of_ribosomal_UMIs_per_cell

Metacells.Contracts.vector_of_std_euclidean_modules_cells_distance_per_metacell

Metacells.Contracts.vector_of_total_UMIs_per_block

Metacells.Contracts.vector_of_total_UMIs_per_cell

Metacells.Contracts.vector_of_total_UMIs_per_metacell

Metacells.Contracts.vector_of_total_neighborhood_UMIs_per_block

Metacells.Contracts.vector_of_type_per_block

Metacells.Contracts.vector_of_type_per_cell

Metacells.Contracts.vector_of_type_per_metacell

Metacells.Gmara.empty_gmara_cache!

Metacells.Gmara.gmara_genes

Metacells.Gmara.normalize_gene_name

Metacells.Gmara.set_gmara_genes_mask!

Metacells.ProjectCells.compute_cells_projection!

Metacells.ProjectCells.compute_matrix_of_correlation_between_neighborhood_cells_and_projected_metacells_per_gene_per_projected_block!

Metacells.ProjectCells.compute_vector_of_correlation_between_cells_and_projected_metacells!

Metacells.SharpenMetacells.sharpen_metacells!