Metacells

Metacells.Metacells — Module

The Metacells.jl package provides computational services for the metacells package, using Daf to hold the data. In the future, we'll ideally migrate all of the metacells package computations to this package, converting the Python package to a thin wrapper, and provide a similar thin R wrapper to provide metacell analysis from R as well. For now, Metacells.jl only provides a subset of the features of the Python metacells package, which requires users to convert data from AnnData (for the old features) to Daf (to the new features).

Index

Metacells.AnalyzeBlocks

Metacells.AnalyzeCells

Metacells.AnalyzeGenes

Metacells.AnalyzeMetacells

Metacells.AnalyzeModules

Metacells.AnnDataFormat

Metacells.ComputeBlocks

Metacells.ComputeModules

Metacells.Contracts

Metacells.Defaults

Metacells.Gmara

Metacells.Metacells

Metacells.SharpenMetacells

Metacells.Defaults.GENE_FRACTION_REGULARIZATION

Metacells.Defaults.MIN_SIGNIFICANT_GENE_UMIS

Metacells.Gmara.CACHE_DIR

Metacells.Gmara.TIMEOUT

Metacells.AnnDataFormat.CopyAnnData

Metacells.AnalyzeBlocks.compute_blocks_covered_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_genes_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_genes_covered_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_is_correlated_with_metacells

Metacells.AnalyzeBlocks.compute_blocks_genes_is_environment_markers!

Metacells.AnalyzeBlocks.compute_blocks_genes_linear_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_log_covered_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_log_linear_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_neighborhood_correlations!

Metacells.AnalyzeBlocks.compute_blocks_max_skeleton_fold_distances!

Metacells.AnalyzeBlocks.compute_blocks_mean_covered_neighborhood_correlations!

Metacells.AnalyzeBlocks.compute_blocks_mean_euclidean_distances!

Metacells.AnalyzeBlocks.compute_blocks_mean_modules_neighborhood_correlations!

Metacells.AnalyzeBlocks.compute_blocks_mean_neighborhood_correlations!

Metacells.AnalyzeBlocks.compute_blocks_n_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_blocks!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_metacells!

Metacells.AnalyzeBlocks.compute_blocks_n_metacells!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_blocks!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_metacells!

Metacells.AnalyzeBlocks.compute_blocks_total_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_types!

Metacells.AnalyzeCells.compute_cells_covered_UMIs!

Metacells.AnalyzeCells.compute_cells_total_UMIs!

Metacells.AnalyzeGenes.fetch_regulators!

Metacells.AnalyzeGenes.identify_covered_genes!

Metacells.AnalyzeGenes.identify_marker_genes!

Metacells.AnalyzeGenes.identify_skeleton_genes!

Metacells.AnalyzeGenes.rank_marker_genes!

Metacells.AnalyzeMetacells.compute_cells_types_by_metacells!

Metacells.AnalyzeMetacells.compute_metacells_covered_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_euclidean_distances!

Metacells.AnalyzeMetacells.compute_metacells_genes_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_genes_covered_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_geomean_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_linear_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_covered_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_geomean_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_linear_fractions!

Metacells.AnalyzeMetacells.compute_metacells_max_skeleton_fold_distances!

Metacells.AnalyzeMetacells.compute_metacells_n_cells!

Metacells.AnalyzeMetacells.compute_metacells_total_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_types_by_cells!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_covered_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_linear_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_covered_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_linear_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_total_UMIs!

Metacells.AnalyzeModules.compute_blocks_modules_is_strong!

Metacells.AnalyzeModules.compute_blocks_modules_n_genes!

Metacells.AnalyzeModules.compute_blocks_modules_n_skeletons!

Metacells.AnnDataFormat.import_cells_h5ad!

Metacells.AnnDataFormat.import_metacells_h5ad!

Metacells.AnnDataFormat.reconstruct_type_axis!

Metacells.ComputeBlocks.compute_blocks!

Metacells.ComputeBlocks.compute_blocks_is_in_environment!

Metacells.ComputeBlocks.compute_blocks_is_in_neighborhood!

Metacells.ComputeModules.compute_blocks_modules!

Metacells.Contracts.block_axis

Metacells.Contracts.block_block_is_in_environment_matrix

Metacells.Contracts.block_block_is_in_neighborhood_matrix

Metacells.Contracts.block_block_max_skeleton_fold_distance

Metacells.Contracts.block_block_mean_euclidean_skeleton_distance

Metacells.Contracts.block_covered_UMIs_vector

Metacells.Contracts.block_gene_UMIs_matrix

Metacells.Contracts.block_gene_covered_fraction_matrix

Metacells.Contracts.block_gene_is_correlated_with_metacells

Metacells.Contracts.block_gene_is_environment_marker_matrix

Metacells.Contracts.block_gene_linear_fraction_matrix

Metacells.Contracts.block_gene_log_covered_fraction_matrix

Metacells.Contracts.block_gene_log_linear_fraction_matrix

Metacells.Contracts.block_gene_module_matrix

Metacells.Contracts.block_gene_neighborhood_correlation_matrix

Metacells.Contracts.block_mean_covered_neighborhood_correlation_vector

Metacells.Contracts.block_mean_modules_neighborhood_correlation_vector

Metacells.Contracts.block_mean_neighborhood_correlation_vector

Metacells.Contracts.block_metacell_module_environment_covered_fraction_tensor

Metacells.Contracts.block_metacell_module_environment_linear_fraction_tensor

Metacells.Contracts.block_metacell_module_environment_log_covered_fraction_tensor

Metacells.Contracts.block_metacell_module_environment_log_linear_fraction_tensor

Metacells.Contracts.block_metacell_module_environment_total_UMIs_tensor

Metacells.Contracts.block_module_is_found_matrix

Metacells.Contracts.block_module_is_strong_matrix

Metacells.Contracts.block_module_n_genes_matrix

Metacells.Contracts.block_module_n_skeletons_matrix

Metacells.Contracts.block_n_cells_vector

Metacells.Contracts.block_n_environment_blocks_vector

Metacells.Contracts.block_n_environment_cells_vector

Metacells.Contracts.block_n_environment_metacells_vector

Metacells.Contracts.block_n_metacells_vector

Metacells.Contracts.block_n_neighborhood_blocks_vector

Metacells.Contracts.block_n_neighborhood_cells_vector

Metacells.Contracts.block_n_neighborhood_metacells_vector

Metacells.Contracts.block_total_UMIs_vector

Metacells.Contracts.block_type_vector

Metacells.Contracts.cell_axis

Metacells.Contracts.cell_covered_UMIs_vector

Metacells.Contracts.cell_excluded_UMIs_vector

Metacells.Contracts.cell_gene_UMIs_matrix

Metacells.Contracts.cell_is_excluded_vector

Metacells.Contracts.cell_metacell_vector

Metacells.Contracts.cell_mitochondrial_UMIs_vector

Metacells.Contracts.cell_ribosomal_UMIs_vector

Metacells.Contracts.cell_total_UMIs_vector

Metacells.Contracts.cell_type_vector

Metacells.Contracts.gene_axis

Metacells.Contracts.gene_is_covered_vector

Metacells.Contracts.gene_is_excluded_vector

Metacells.Contracts.gene_is_forbidden_vector

Metacells.Contracts.gene_is_lateral_vector

Metacells.Contracts.gene_is_marker_vector

Metacells.Contracts.gene_is_mitochondrial_vector

Metacells.Contracts.gene_is_regulator_vector

Metacells.Contracts.gene_is_ribosomal_vector

Metacells.Contracts.gene_is_skeleton_vector

Metacells.Contracts.gene_marker_rank_vector

Metacells.Contracts.metacell_axis

Metacells.Contracts.metacell_block_vector

Metacells.Contracts.metacell_covered_UMIs_vector

Metacells.Contracts.metacell_gene_UMIs_matrix

Metacells.Contracts.metacell_gene_covered_fraction_matrix

Metacells.Contracts.metacell_gene_geomean_fraction_matrix

Metacells.Contracts.metacell_gene_linear_fraction_matrix

Metacells.Contracts.metacell_gene_log_covered_fraction_matrix

Metacells.Contracts.metacell_gene_log_geomean_fraction_matrix

Metacells.Contracts.metacell_gene_log_linear_fraction_matrix

Metacells.Contracts.metacell_metacell_euclidean_skeleton_distance

Metacells.Contracts.metacell_metacell_max_skeleton_fold_distance

Metacells.Contracts.metacell_n_cells_vector

Metacells.Contracts.metacell_total_UMIs_vector

Metacells.Contracts.metacell_type_vector

Metacells.Contracts.module_anchor_gene_vector

Metacells.Contracts.module_axis

Metacells.Contracts.type_axis

Metacells.Contracts.type_color_vector

Metacells.Gmara.empty_gmara_cache!

Metacells.Gmara.gmara_genes

Metacells.Gmara.normalize_gene_name

Metacells.Gmara.set_gmara_genes_mask!

Metacells.SharpenMetacells.sharpen_metacells!