Metacells

Metacells.Metacells — Module

The Metacells.jl package provides computational services for the metacells package, using Daf to hold the data. In the future, we'll ideally migrate all of the metacells package computations to this package, converting the Python package to a thin wrapper, and provide a similar thin R wrapper to provide metacell analysis from R as well. For now, Metacells.jl only provides a subset of the features of the Python metacells package, which requires users to convert data from AnnData (for the old features) to Daf (to the new features).

Index

Metacells.AnnDataFormat

Metacells.Boxes

Metacells.Contracts

Metacells.Defaults

Metacells.Gmara

Metacells.IdentifyGenes

Metacells.Metacells

Metacells.Defaults.GENE_FRACTION_REGULARIZATION

Metacells.Defaults.MIN_SIGNIFICANT_GENE_UMIS

Metacells.Gmara.CACHE_DIR

Metacells.Gmara.TIMEOUT

Metacells.AnnDataFormat.CopyAnnData

Metacells.AnnDataFormat.import_h5ads!

Metacells.Boxes.compute_boxes!

Metacells.Boxes.compute_boxes_data!

Metacells.Contracts.box_axis

Metacells.Contracts.box_box_distance

Metacells.Contracts.box_main_neighborhood_vector

Metacells.Contracts.box_neighborhood_is_member_matrix

Metacells.Contracts.box_total_UMIs_vector

Metacells.Contracts.box_type_vector

Metacells.Contracts.cell_axis

Metacells.Contracts.cell_is_excluded_vector

Metacells.Contracts.gene_axis

Metacells.Contracts.gene_box_fraction_matrix

Metacells.Contracts.gene_box_total_UMIs_matrix

Metacells.Contracts.gene_divergence_vector

Metacells.Contracts.gene_is_correlated_vector

Metacells.Contracts.gene_is_excluded_vector

Metacells.Contracts.gene_is_lateral_vector

Metacells.Contracts.gene_is_marker_vector

Metacells.Contracts.gene_metacell_fraction_matrix

Metacells.Contracts.gene_metacell_total_UMIs_matrix

Metacells.Contracts.gene_neighborhood_is_correlated_matrix

Metacells.Contracts.metacell_axis

Metacells.Contracts.metacell_box_vector

Metacells.Contracts.metacell_total_UMIs_vector

Metacells.Contracts.metacell_type_vector

Metacells.Contracts.neighborhood_axis

Metacells.Contracts.neighborhood_span_vector

Metacells.Contracts.type_axis

Metacells.Contracts.type_color_vector

Metacells.Gmara.empty_gmara_cache!

Metacells.Gmara.gmara_genes

Metacells.Gmara.normalize_gene_name

Metacells.Gmara.set_gmara_genes_mask!

Metacells.IdentifyGenes.compute_genes_divergence!

Metacells.IdentifyGenes.identify_correlated_genes!

Metacells.IdentifyGenes.identify_marker_genes!