Analyze Blocks

Metacells.AnalyzeBlocks — Module

Do simple blocks analysis.

Counts

Cells

Metacells.AnalyzeBlocks.compute_blocks_n_cells! — Function

function compute_blocks_n_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of cells per block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each metacell.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Outputs

Vectors

block @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of cells in the metacells in each block.

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_cells! — Function

function compute_blocks_n_neighborhood_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of cells in the metacells of the blocks of the neighborhood centered at a block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each metacell.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

block, block @ is in neighborhood ::Bool (required): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of cells in the metacells in the blocks of the neighborhood centered at each block.

Metacells.AnalyzeBlocks.compute_blocks_n_environment_cells! — Function

function compute_blocks_n_environment_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of cells in the metacells of the blocks of the environment centered at each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each metacell.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

Outputs

Vectors

block @ n environment cells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of cells in the metacells in the blocks of the environment centered at each block.

Metacells

Metacells.AnalyzeBlocks.compute_blocks_n_metacells! — Function

function compute_blocks_n_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of metacells per block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Outputs

Vectors

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of metacells in each block.

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_metacells! — Function

function compute_blocks_n_neighborhood_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of metacells in the blocks of the neighborhood centered at each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

block, block @ is in neighborhood ::Bool (required): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood metacells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of metacells in the blocks of the neighborhood centered at each block.

Metacells.AnalyzeBlocks.compute_blocks_n_environment_metacells! — Function

function compute_blocks_n_environment_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of metacells in the blocks of the environment centered at each block.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

Outputs

Vectors

block @ n environment metacells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of metacells in the blocks of the environment centered at each block.

Blocks

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_blocks! — Function

function compute_blocks_n_neighborhood_blocks!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of blocks in the neighborhood centered at each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

block, block @ is in neighborhood ::Bool (required): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

Outputs

Vectors

block @ n neighborhood blocks ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of blocks in the neighborhood centered at each block.

Metacells.AnalyzeBlocks.compute_blocks_n_environment_blocks! — Function

function compute_blocks_n_environment_blocks!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of blocks in the environment centered at each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

Outputs

Vectors

block @ n environment blocks ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of blocks in the environment centered at each block.

UMIs

Metacells.AnalyzeBlocks.compute_blocks_genes_UMIs! — Function

function compute_blocks_genes_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of UMIs used to estimate the fraction of each gene in each block. This can used to estimate the robustness of the fraction.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of each gene in each block.

Metacells.AnalyzeBlocks.compute_blocks_total_UMIs! — Function

function compute_blocks_total_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of genes per block.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Vectors

block @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of all the non-excluded genes in each block.

Metacells.AnalyzeBlocks.compute_blocks_covered_UMIs! — Function

function compute_blocks_covered_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of covered genes per block.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Vectors

block @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of the all covered genes in each block.

Fractions

Metacells.AnalyzeBlocks.compute_blocks_genes_linear_fractions! — Function

function compute_blocks_genes_linear_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

An estimated linear fraction of the UMIs of each gene in each block. This is just the total UMIs of the gene in the metacell divided by the total UMIs of the metacell, which is the "best" estimate assuming multinomial sampling noise. However, this is sensitive to a few metacells with very high expression levels ("bursty" genes).

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Matrices

gene, block @ linear_fraction ::Union{Float32, Float64} (guaranteed): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Metacells.AnalyzeBlocks.compute_blocks_genes_log_linear_fractions! — Function

function compute_blocks_genes_log_linear_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of each gene in each block. This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

gene, block @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Outputs

Matrices

gene, block @ log linear fraction ::Union{Float32, Float64} (guaranteed): The log base 2 of the linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Metacells.AnalyzeBlocks.compute_blocks_genes_covered_fractions! — Function

function compute_blocks_genes_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

An estimated linear fraction of the UMIs of each covered gene in each block. By considering only the covered genes this avoid the impact of highly-expressed lateral genes (e.g., cell cycle). Otherwise is similar to compute_blocks_genes_linear_fractions! .

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

Matrices

gene, block @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in each block.

Outputs

Matrices

gene, block @ covered_fraction ::Union{Float32, Float64} (guaranteed): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Metacells.AnalyzeBlocks.compute_blocks_genes_log_covered_fractions! — Function

function compute_blocks_genes_log_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of each covered gene in each block. This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

block (required): Distinct groups of metacells with "very close" cell state.

Matrices

gene, block @ covered_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene out of the total UMIs in each block.

Outputs

Matrices

gene, block @ log covered fraction ::Union{Float32, Float64} (guaranteed): The log base 2 of the linear fraction of the UMIs of each covered gene out of the total UMIs in each block.

Genes

Metacells.AnalyzeBlocks.compute_blocks_genes_is_environment_markers! — Function

compute_blocks_genes_is_environment_markers!(
    daf::DafWriter;
    min_marker_gene_max_fraction::AbstractFloat = ```0.0001```,
    min_marker_gene_range_fold::Real = ```2```,
    overwrite::Bool = false,
)::Nothing

A mask of genes that distinguish between cell states in each environment. Otherwise similar to identify_marker_genes! .

Metadata

Metacells.AnalyzeBlocks.compute_blocks_types! — Function

function compute_blocks_types!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The type of each block, based on is metacell types.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

metacell @ type ::AbstractString (required): The type each metacell belongs to.

Outputs

Vectors

block @ type ::AbstractString (guaranteed): The type each block belongs to.

Distances

Metacells.AnalyzeBlocks.compute_blocks_mean_euclidean_distances! — Function

compute_blocks_mean_euclidean_distances!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The mean Euclidean skeleton genes covered fractions distance between the metacells of the blocks.

Metacells.AnalyzeBlocks.compute_blocks_max_skeleton_fold_distances! — Function

compute_blocks_max_skeleton_fold_distances!(
    daf::DafWriter;
    overwrite::Bool = false,
)::Nothing

The maximal significant skeleton genes covered fractions fold factor between metacells of the blocks.

Index

Metacells.AnalyzeBlocks

Metacells.AnalyzeBlocks.compute_blocks_covered_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_genes_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_genes_covered_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_is_environment_markers!

Metacells.AnalyzeBlocks.compute_blocks_genes_linear_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_log_covered_fractions!

Metacells.AnalyzeBlocks.compute_blocks_genes_log_linear_fractions!

Metacells.AnalyzeBlocks.compute_blocks_max_skeleton_fold_distances!

Metacells.AnalyzeBlocks.compute_blocks_mean_euclidean_distances!

Metacells.AnalyzeBlocks.compute_blocks_n_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_blocks!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_environment_metacells!

Metacells.AnalyzeBlocks.compute_blocks_n_metacells!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_blocks!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_cells!

Metacells.AnalyzeBlocks.compute_blocks_n_neighborhood_metacells!

Metacells.AnalyzeBlocks.compute_blocks_total_UMIs!

Metacells.AnalyzeBlocks.compute_blocks_types!