Analyze Genes

Metacells.AnalyzeGenes — Module

Do simple per-gene analysis.

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_regulator_per_gene! — Function

fetch_gmara_vector_of_is_regulator_per_gene!(
    daf::DafWriter;
    species::AbstractString,
    namespace::AbstractString = ```"GeneSymbol"```,
    version::AbstractString = ```"main"```,
    cache_dir = ```"/Users/obk/.cach..." (23)```,
    timeout::Real = ```10```,
    overwrite::Bool = ```false```,
)::Nothing

Fetch and set vector_of_is_transcription_factor_per_gene from Gmara .

Inputs

Axes

gene (required): Sequenced genes.

Outputs

Vectors

gene @ is_regulator ::Bool (created): A mask of genes that regulate the expression level of other genes.

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_transcription_factor_per_gene! — Function

fetch_gmara_transcription_factors!(
    daf::DafWriter;
    species::AbstractString,
    namespace::AbstractString = ```"GeneSymbol"```,
    version::AbstractString = ```"main"```,
    cache_dir = ```"/Users/obk/.cach..." (23)```,
    timeout::Real = ```10```,
    overwrite::Bool = ```false```,
)::Nothing

Fetch and set vector_of_is_transcription_factor_per_gene from Gmara .

Inputs

Axes

gene (required): Sequenced genes.

Outputs

Vectors

gene @ is transcription factor ::Bool (created): A mask of genes that regulate the expression level of other genes.

Metacells.AnalyzeGenes.compute_vector_of_is_marker_per_gene! — Function

function compute_vector_of_is_marker_per_gene!(
    daf::DafWriter;
    min_marker_gene_max_fraction::AbstractFloat = ```0.0001```,
    min_marker_gene_range_fold::Real = ```2```,
    min_marker_quantile::Real = ```0```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_is_marker_per_gene . Such genes:

Have a maximal gene expression level of at least min_marker_gene_max_fraction , and

Have a range of expression of at least min_marker_gene_range_fold between the maximal and minimal (log base 2) of the expression level. For the minimal expression, we take the min_marker_quantile of the expression levels in the population.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ is_marker ::Bool (created): A mask of genes that distinguish between cell states.

Metacells.AnalyzeGenes.compute_vector_of_marker_rank_per_gene! — Function

function compute_vector_of_marker_rank_per_gene!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_marker_rank_per_gene . To compute this:

Non-marker genes are given a rank of typemax(UInt32) regardless of their expression level in the metacells (see vector_of_is_marker_per_gene ).

Compute the median of the (log base 2) of the gene expression level for each marker gene across all the metacells.

Compute the per-marker-per-metacell fold factor (absolute difference of the log expression from the median).

Invoke rank_variables to rank the marker genes from most to least significant.

Non-marker genes are given a rank of typemax(UInt32) regardless of their expression level in the metacells.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ marker_rank ::Union{UInt16, UInt32, UInt64, UInt8} (created): The ralative ranks of the marker genes.

Metacells.AnalyzeGenes.compute_vector_of_is_skeleton_per_gene! — Function

function compute_is_skeleton_per_gene_vector!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_is_skeleton_per_gene . We just pick the genes that are regulators, and are not excluded, lateral or forbidden.

Inputs

Axes

gene (required): Sequenced genes.

Vectors

gene @ is_regulator ::Bool (required): A mask of genes that regulate the expression level of other genes.

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

gene @ is_forbidden ::Bool (required): A mask of genes that are forbidden from being used as skeleton genes.

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

Outputs

Vectors

gene @ is_skeleton ::Bool (created): A mask of genes that are used to predict the values of the rest of the genes.

Metacells.AnalyzeGenes.compute_vector_of_is_correlated_with_skeleton_per_gene! — Function

function compute_vector_of_is_correlated_with_skeleton_per_gene!(
    daf::DafWriter;
    min_gene_correlation::AbstractFloat = ```0.6```,
    min_gene_correlation_quantile::AbstractFloat = ```0.9```,
    genes_correlation_window::Integer = ```100```,
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_is_correlated_with_skeleton_per_gene . Such genes:

Are markers (see vector_of_is_marker_per_gene .

Have a correlation of at least min_gene_correlation with some skeleton gene (see vector_of_is_skeleton_per_gene ), OR

Have a correlation with a skeleton gene which is in the min_gene_correlation_quantile out of the genes_correlation_window genes with a similar correlation with some skeleton.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the genes.

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ is correlated with_skeleton ::Bool (created): A mask of (marker) genes that have significant correlation with any of the skeleton genes.

Metacells.AnalyzeGenes.rank_variables — Function

rank_variables(score_per_variable_per_observation::AbstractMatrix{<:AbstractFloat})::AbstractVector{<:Integer}

Given some score_per_variable_per_observation matrix, return a vector of the rank of each variable such that the "most significant" variables are first.

Rank the variables for each observation (1 having the highest score).

For each variable, give it a priority which is a tuple of (1) the minimal rank it has in all observations (2) the maximal score it has in observations where it has that rank (negated).

Sort the variables according to this priority.

This heuristic is useful for focusing on the "most significant" variables in a data set. It is used by compute_vector_of_marker_rank_per_gene! .

TODO: Move to TanayLabUtilities ?

Index

Metacells.AnalyzeGenes

Metacells.AnalyzeGenes.compute_vector_of_is_correlated_with_skeleton_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_is_marker_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_is_skeleton_per_gene!

Metacells.AnalyzeGenes.compute_vector_of_marker_rank_per_gene!

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_regulator_per_gene!

Metacells.AnalyzeGenes.fetch_gmara_vector_of_is_transcription_factor_per_gene!

Metacells.AnalyzeGenes.rank_variables