Analyze Genes

Metacells.AnalyzeGenes — Module

Do simple per-gene analysis.

Metacells.AnalyzeGenes.identify_marker_genes! — Function

function identify_marker_genes!(
    daf::DafWriter;
    min_marker_gene_max_fraction::AbstractFloat = ```0.0001```,
    min_marker_gene_range_fold::Real = ```2```,
    overwrite::Bool = ```false```,
)::Nothing

Identify the genes that distinguish at least one metacell from the rest. Such genes are called "marker" genes as they (potentially) mark specific cell states. If overwrite , will overwrite an existing is_marker mask.

Marker genes are genes which:

Have a maximal gene expression level of at least min_marker_gene_max_fraction , and

Have a range of expression of at least min_marker_gene_range_fold between the maximal and minimal (log base 2) of the expression level.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ is_marker ::Bool (guaranteed): A mask of genes that distinguish between cell states.

Metacells.AnalyzeGenes.rank_marker_genes! — Function

function rank_marker_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute the relative ranks of marker genes.

Compute the median of the (log base 2) of the gene expression level for each marker gene across all the metacells.

Compute the per-marker-per-metacell fold factor (absolute difference of the log expression from the median).

Rank the markers for each metacell (1 having the largest fold factor relative to the median).

For each marker, give it a priority which is a tuple of (1) the minimal rank it has in all metacells (2) the maximal fold it has in metacells where it has that rank (negated).

Sort the markers according to this priority.

Non-marker genes are given a rank of typemax(UInt32) regardless of their expression level in the metacells.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Vectors

gene @ marker_rank ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The ralative ranks of the marker genes.

Metacells.AnalyzeGenes.identify_covered_genes! — Function

function identify_covered_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Identify the genes that will be approximated by the local linear programs. Picking them is simple: we cover all marker genes, that are not lateral.

Inputs

Axes

gene (required): Sequenced genes.

Vectors

gene @ is_marker ::Bool (required): A mask of genes that distinguish between cell states.

gene @ is_lateral ::Bool (required): A mask of genes that are lateral to the biological behaviors of interest.

Outputs

Vectors

gene @ is_covered ::Bool (guaranteed): A mask of genes that are covered by the model(s).

Metacells.AnalyzeGenes.identify_skeleton_genes! — Function

function identify_skeleton_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Identify the skeleton genes that will be used to predict the rest of the (covered) genes. We just pick the covered genes that are also regulators, and are not forbidden.

Inputs

Axes

gene (required): Sequenced genes.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

gene @ is_regulator ::Bool (required): A mask of genes that regulate the expression level of other genes.

gene @ is_forbidden ::Bool (required): A mask of genes that are forbidden from being used as skeleton genes.

Outputs

Vectors

gene @ is_skeleton ::Bool (guaranteed): A mask of genes that are used to predict the values of the rest of the (covered) genes.

Metacells.AnalyzeGenes.fetch_regulators! — Function

fetch_regulators!(
    daf::DafWriter;
    species::AbstractString,
    namespace::AbstractString = ```"GeneSymbol"```,
    version::AbstractString = ```"main"```,
    cache_dir = ```"/Users/obk/.cach..." (23)```,
    timeout::Real = ```10```,
    overwrite::Bool = ```false```,
)::Nothing

Fetch the names of transcription factor genes from Gmara and create a mask based on them. These genes will be assumed to be the core genes driving gene programs that describe cell behavior.

Index

Metacells.AnalyzeGenes

Metacells.AnalyzeGenes.fetch_regulators!

Metacells.AnalyzeGenes.identify_covered_genes!

Metacells.AnalyzeGenes.identify_marker_genes!

Metacells.AnalyzeGenes.identify_skeleton_genes!

Metacells.AnalyzeGenes.rank_marker_genes!