Analyze Metacells

Metacells.AnalyzeMetacells — Module

Do simple metacells analysis.

Counts

Metacells.AnalyzeMetacells.compute_metacells_n_cells! — Function

function compute_metacells_n_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of metacells per metacell.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Outputs

Vectors

metacell @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of cells in each metacell.

UMIs

Metacells.AnalyzeMetacells.compute_metacells_genes_UMIs! — Function

function compute_metacells_genes_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of UMIs used to estimate the fraction of each gene in each metacell.

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of each gene in the cells of each metacell.

Metacells.AnalyzeMetacells.compute_metacells_total_UMIs! — Function

function compute_metacells_total_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total number of UMIs used to estimate the fraction of all the genes in each metacell.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of UMIs of all the non-excluded genes in each metacell.

Metacells.AnalyzeMetacells.compute_metacells_covered_UMIs! — Function

function compute_metacells_covered_UMIs!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The total UMIs of covered genes per metacell.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Vectors

metacell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The total number of the covered genes in each metacell.

Linear Fractions

Metacells.AnalyzeMetacells.compute_metacells_genes_linear_fractions! — Function

function compute_metacells_genes_linear_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

An estimated linear fraction of the UMIs of each gene in each metacell. This is just the total UMIs of the gene in the metacell divided by the total UMIs of the metacell, which is the "best" estimate assuming multinomial sampling noise. However, this is sensitive to a few cells with very high expression levels ("bursty" genes).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (guaranteed): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Metacells.AnalyzeMetacells.compute_metacells_genes_log_linear_fractions! — Function

function compute_metacells_genes_log_linear_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of each gene in each metacell. This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ linear_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Outputs

Matrices

gene, metacell @ log linear fraction ::Union{Float32, Float64} (guaranteed): The log base 2 of the linear fraction of the UMIs of each non-excluded gene in each metacell, out of the total UMIs.

Covered Fractions

Metacells.AnalyzeMetacells.compute_metacells_genes_covered_fractions! — Function

function compute_metacells_genes_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

An estimated linear fraction of the UMIs of each covered gene in each metacell. By considering only the covered genes this avoid the impact of highly-expressed lateral genes (e.g., cell cycle). Otherwise is similar to compute_metacells_genes_linear_fractions! .

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the model(s).

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (guaranteed): The linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

Metacells.AnalyzeMetacells.compute_metacells_genes_log_covered_fractions! — Function

function compute_metacells_genes_log_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the estimated linear fraction of the UMIs of each covered gene in each metacell. This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

Outputs

Matrices

metacell, gene @ log covered fraction ::Union{Float32, Float64} (guaranteed): The log base 2 of the linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

Geomean Fractions

Metacells.AnalyzeMetacells.compute_metacells_genes_geomean_fractions! — Function

compute_metacells_genes_geomean_fractions!(
    daf::DafWriter;
    UMIs_regularization::AbstractFloat = ```0.0625```,
    min_downsamples::Integer = ```750```,
    min_downsamples_quantile::AbstractFloat = ```0.05```,
    max_downsamples_quantile::AbstractFloat = ```0.5```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Given an assignment of cells to metacell, compute an geomean estimation of the fraction of UMIs of each gene for each metacell.

The linear way to do this would be to just take the total UMIs of the gene out of the total UMIs of the metacell. However, this method has a weakness; a single strong cell with a "very different" fraction of the gene will dominate (that is, the method is sensitive to outliers).

Instead, we take the geometric mean of the fractions of the gene in the cells. This however raises a few issues we need to deal with:

Genes with zero UMIs are a problem; to combat this, we add a UMIs_regularization factor when computing the fractions, take the geomean, and subtract the regularization at the end (so all-zero genes will still get a zero overall fraction).

We want to give more weight to cells with more UMIs, so we use a scaled geomean. The weight we give to each cell is the log of the total number of UMIs in it.

The geomean fractions of all the genes in a metacell do not sum to one, so we scale them. This has the unfortunate side effect that the end result does not obey a nice relation to the linear fraction of the gene in each of the cells; in particular, the final result might be higher than all of these per-cell fractions (ouch).

This raises an important point about the whole "fraction of UMIs of a gene in a cell" concept, which is that it is highly dependent on the set of genes you pick to compute the fraction out of (that is, the denominator). This places restrictions on how you should use these fractions:

You can't just compare these fractions between two arbitrary data sets, as the denominators aren't the same. You must identify the set of common genes, and renormalize the fractions to sum to one in this subset, in both data sets. Only then can you meaningfully compare the results.

Even if you compare two data sets with the same set of genes (or even two subsets of the same data set, such as two "cell types"), if a specific "gene program" has a very high total expression in only one of them, then all other genes will artificially appear to be lower.

This is why we recommend excluding ribosomal genes from the data sets; they can take up anything between almost none to over two thirds of the total UMIs, and this varies between "cell types". This means that if they are included in the denominator, the fractions of otherwise "identical" genes will appear to differ by a factor of up to 3X

Luckily, most gene programs we deal with have a total expression of a few percent at most, so this effect is negligible. However if you identify a gene program with a total expression higher than, say, 10%, you should consider its effect on the rest of the genes.

Inputs

Axes

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_excluded ::Bool (required): A mask of genes that are excluded from consideration.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Matrices

gene, metacell @ geomean_fraction ::Union{Float32, Float64} (guaranteed): The geomean fraction of the UMIs of each gene in each metacell.

Metacells.AnalyzeMetacells.compute_metacells_genes_log_geomean_fractions! — Function

function compute_metacells_genes_log_geomean_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the estimated geomean fraction of the UMIs of each gene in each metacell. This adds the gene_fraction_regularization to deal with zero fractions.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

gene, metacell @ geomean_fraction ::Union{Float32, Float64} (required): The geomean fraction of the UMIs of each gene in each metacell.

Outputs

Matrices

gene, metacell @ log geomean fraction ::Union{Float32, Float64} (guaranteed): The log base 2 of the geomean fraction of the UMIs of each gene in each metacell.

Types

Metacells.AnalyzeMetacells.compute_cells_types_by_metacells! — Function

function compute_cells_types_by_metacells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The type of each cell, based on the type of the metacell it belongs to.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ type ::AbstractString (required): The type each metacell belongs to.

Outputs

Vectors

cell @ type ::AbstractString (guaranteed): The type each cell belongs to.

Metacells.AnalyzeMetacells.compute_metacells_types_by_cells! — Function

function compute_metacells_types_by_cells!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The type of each metacell, based on is cell types. This assumes that each cell was assigned type, either because we imported the data from a cell-based atlas, or from previous computed metacells.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

cell @ type ::AbstractString (required): The type each cell belongs to.

Outputs

Vectors

metacell @ type ::AbstractString (guaranteed): The type each metacell belongs to.

Distances

Metacells.AnalyzeMetacells.compute_metacells_euclidean_distances! — Function

function compute_metacells_euclidean_distances!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The Euclidean distance between the log of the covered fraction of the skeleton genes between the metacells.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

Matrices

metacell, gene @ log covered fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

Outputs

Matrices

metacell, metacell @ euclidean skeleton distance ::Union{Float32, Float64} (guaranteed): The Euclidean distance between the log of the covered fraction of the skeleton genes between the metacells.

Metacells.AnalyzeMetacells.compute_metacells_max_skeleton_fold_distances! — Function

function compute_metacells_max_skeleton_fold_distances!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    min_significant_gene_UMIs::Integer = ```40```,
    fold_confidence::AbstractFloat = ```0.9```,
    overwrite::Bool = ```false```,
)::Nothing

The maximal significant fold factor between the covered fraction of skeleton genes between the metacells. This uses heuristics to require the fold factor be based on a sufficient number of UMIs to be robust.

The fold factor is log (base 2) of the gene expression using the gene_fraction_regularization . For computing this fold factor, we ignore "insignificant" genes whose total UMIs in the compared metacells isn't at least min_significant_gene_UMIs . We also we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells.

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

gene (required): Sequenced genes.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

metacell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of the covered genes in each metacell.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): The linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

Outputs

Matrices

metacell, metacell @ max skeleton fold_distance ::Union{Float32, Float64} (guaranteed): The maximal significant fold factor between the covered fraction of skeleton genes between the metacells.

Index

Metacells.AnalyzeMetacells

Metacells.AnalyzeMetacells.compute_cells_types_by_metacells!

Metacells.AnalyzeMetacells.compute_metacells_covered_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_euclidean_distances!

Metacells.AnalyzeMetacells.compute_metacells_genes_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_genes_covered_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_geomean_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_linear_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_covered_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_geomean_fractions!

Metacells.AnalyzeMetacells.compute_metacells_genes_log_linear_fractions!

Metacells.AnalyzeMetacells.compute_metacells_max_skeleton_fold_distances!

Metacells.AnalyzeMetacells.compute_metacells_n_cells!

Metacells.AnalyzeMetacells.compute_metacells_total_UMIs!

Metacells.AnalyzeMetacells.compute_metacells_types_by_cells!