Analyze Modules

Metacells.AnalyzeModules — Module

Do simple gene module analysis.

Metacells.AnalyzeModules.compute_vector_of_n_modules_per_block! — Function

function compute_vector_of_n_modules_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_n_modules_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

Matrices

block, module @ is_found ::Bool (required): A mask of the modules that were found for each block.

Outputs

Vectors

block @ n_modules ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of found modules in each block.

Metacells.AnalyzeModules.compute_matrix_of_n_genes_per_module_per_block! — Function

function compute_matrix_of_n_genes_per_module_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_n_genes_per_module_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

gene (required): Sequenced genes.

Matrices

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

Outputs

Matrices

block, module @ n_genes ::Union{UInt16, UInt32, UInt64, UInt8} (created): The number of genes in each module in each block.

Metacells.AnalyzeModules.compute_stats_of_euclidean_modules_cells_distance_per_metacell! — Function

compute_stats_of_euclidean_modules_cells_distance_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set vector_of_mean_euclidean_modules_cells_distance_per_metacell and and set vector_of_std_euclidean_modules_cells_distance_per_metacell .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

gene (required): Sequenced genes.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

Matrices

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

Outputs

Vectors

metacell @ mean euclidean modules cells distance ::Union{Float32, Float64} (created): In each metacell, the mean of the euclidean distances of between its cells and the metacell's linear fraction of the modules of the metacell's block's neighborhood.

metacell @ std euclidean modules cells distance ::Union{Float32, Float64} (created): In each metacell, the standard deviation of the euclidean distances of between its cells and the metacell's linear fraction of the modules of the metacell's block's neighborhood.

Metacells.AnalyzeModules.compute_stats_of_linear_fraction_in_neighborhood_cells_per_module_per_block! — Function

compute_stats_of_linear_fraction_in_neighborhood_cells_per_module_per_block!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set matrix_of_mean_linear_fraction_in_neighborhood_cells_per_module_per_block and and set matrix_of_std_linear_fraction_in_neighborhood_cells_per_module_per_block .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

cell (required): Sequenced single cells.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

gene (required): Sequenced genes.

Vectors

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

Outputs

Matrices

block, module @ mean linear fraction in neighborhood_cells ::Union{Float32, Float64} (created): The mean of the linear fraction of each module in the cells of the neighborhood of each block.

block, module @ std linear fraction in neighborhood_cells ::Union{Float32, Float64} (created): The standard deviation of the linear fraction of each module in the cells of the neighborhood of each block.

Metacells.AnalyzeModules.compute_tensor_of_linear_fraction_per_block_per_module_per_metacell! — Function

compute_tensor_of_linear_fraction_per_block_per_module_per_metacell!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

Compute and set compute_tensor_of_linear_fraction_per_block_per_module_per_metacell! .

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

metacell (required): Minimal-sized groups of cells for robust point estimates.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

gene (required): Sequenced genes.

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each metacell.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of each gene in the cells of each metacell.

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

Outputs

Tensors

block; metacell, module @ linear_fraction ::Union{Float32, Float64} (created): The linear fraction of the UMIs of each block's found module genes out of the total UMIs in each metacell.

Index

Metacells.AnalyzeModules

Metacells.AnalyzeModules.compute_matrix_of_n_genes_per_module_per_block!

Metacells.AnalyzeModules.compute_stats_of_euclidean_modules_cells_distance_per_metacell!

Metacells.AnalyzeModules.compute_stats_of_linear_fraction_in_neighborhood_cells_per_module_per_block!

Metacells.AnalyzeModules.compute_tensor_of_linear_fraction_per_block_per_module_per_metacell!

Metacells.AnalyzeModules.compute_vector_of_n_modules_per_block!