Analyze Modules

Metacells.AnalyzeModules — Module

Do simple gene module analysis.

Metacells.AnalyzeModules.compute_blocks_modules_is_strong! — Function

function compute_blocks_modules_is_strong!(
    daf::DafWriter;
    min_downsamples::Integer = ```750```,
    min_downsamples_quantile::AbstractFloat = ```0.05```,
    max_downsamples_quantile::AbstractFloat = ```0.5```,
    min_strong_UMIs::Integer = 8,
    min_strong_cells::Integer = ```12```,
    overwrite::Bool = ```false```,
)::Nothing

A mask of the strong modules that have enough UMIs in enough cells. Enough UMIs means that, if we downsample the cells to a common total covered UMIs using downsamples with the min_downsamples , min_downsamples_quantile , max_downsamples_quantile , the expected total UMIs of the module will be at least min_strong_UMIs . Enough cells mean that this will happen in at least min_strong_cells in the environment. Modules which achieve this are marked as strong enough to be meaningfully applicable to single cells.

Inputs

Axes

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

gene (required): Sequenced genes.

module (required): Groups of genes that together predict the rest of the covered genes.

Vectors

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

cell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the covered genes in each cell.

Matrices

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

block, module @ is_found ::Bool (required): A mask of the modules that were found for each block.

Outputs

Matrices

block, module @ is_strong ::Bool (guaranteed): A mask of the modules that have enough UMIs in enough cells.

Metacells.AnalyzeModules.compute_blocks_modules_n_genes! — Function

function compute_blocks_modules_n_genes!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of genes in each gene module in each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

module (required): Groups of genes that together predict the rest of the covered genes.

gene (required): Sequenced genes.

Matrices

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

Outputs

Matrices

block, module @ n_genes ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of genes in each gene module in each block.

Metacells.AnalyzeModules.compute_blocks_modules_n_skeletons! — Function

function compute_blocks_modules_n_skeletons!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The number of skeleton genes in each gene module in each block.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

module (required): Groups of genes that together predict the rest of the covered genes.

gene (required): Sequenced genes.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

Matrices

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

Outputs

Matrices

block, module @ n_skeletons ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of skeleton genes in each gene module in each block.

Metacells.AnalyzeModules.compute_block_metacell_module_environment_total_UMIs! — Function

compute_block_metacell_module_environment_total_UMIs!(daf::DafWriter; overwrite::Bool = false)::Nothing

The total UMIs of each environment module in each environment metacell.

Metacells.AnalyzeModules.compute_block_metacell_module_environment_linear_fractions! — Function

compute_block_metacell_module_environment_linear_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```
)::Nothing

The linear fraction of the total UMIs of each environment module in each environment metacell, out of the total UMIs.

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_linear_fractions! — Function

compute_block_metacell_module_environment_log_linear_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the linear fraction of the total UMIs of each environment module in each environment metacell, out of the total UMIs.

Metacells.AnalyzeModules.compute_block_metacell_module_environment_covered_fractions! — Function

compute_block_metacell_module_environment_covered_fractions!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

The linear fraction of the total UMIs of each environment module in each environment metacell, out of the total covered UMIs.

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_covered_fractions! — Function

compute_block_metacell_module_environment_log_covered_fractions!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    overwrite::Bool = ```false```,
)::Nothing

The log base 2 of the linear fraction of the total UMIs of each environment module in each environment metacell, out of the total covered UMIs.

Index

Metacells.AnalyzeModules

Metacells.AnalyzeModules.compute_block_metacell_module_environment_covered_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_linear_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_covered_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_log_linear_fractions!

Metacells.AnalyzeModules.compute_block_metacell_module_environment_total_UMIs!

Metacells.AnalyzeModules.compute_blocks_modules_is_strong!

Metacells.AnalyzeModules.compute_blocks_modules_n_genes!

Metacells.AnalyzeModules.compute_blocks_modules_n_skeletons!