AnnData Format

Metacells.AnnDataFormat — Module

Import and export metacells data from/to h5ad files. This allows moving data between the old Python/C++ based AnnData world and the new Julia based Daf world.

Metacells.AnnDataFormat.import_h5ads! — Function

function import_h5ads!(;
    destination::DafWriter,
    raw_cells_h5ad::Maybe{AbstractString} = nothing,
    clean_cells_h5ad::AbstractString,
    metacells_h5ad::AbstractString,
    type_property::Maybe{AbstractString} = nothing,
    rename_type::Maybe{AbstractString} = "type",
    type_colors_csv::Maybe{AbstractString} = nothing,
    type_properties::Maybe{AbstractStringSet} = nothing,
    properties_defaults::Maybe{Dict} = nothing,
)::Nothing

Import an AnnData based metacells dataset into a Daf destination data set. Ideally, the input must include clean_cells_h5ad and the metacells_h5ad computed for them, and optionally also the raw_cells_h5ad including the excluded cells and genes.

If type annotations were assigned to the metacells, then the name of the type_property should be specified. This can be further enhanced by specifying a type_colors_csv file mapping type names to colors. This should be a comma or tab separated file containing at least two columns, one named "color" and one with the same name as the type_property . For consistency, by default the type_property is renamed to the value of rename_type (by default, "type"). You can disable this by setting rename_type to nothing . We also call reconstruct_axis! to build the type axis; you can therefore specify an empty_type name, which will be converted to the empty string, to match the Daf convention of "no value" for string data, and specify an explicit set of type_properties (by default, any per-metacell property that has the same value for all metacells of each type will be converted to a type property) and properties_defaults .

This will mostly just read all the specified h5ad files and copy the data into the destination , with the following changes to match the $Daf$ capabilities and conventions:

The X matrix of the cells is renamed to UMIs , and the X matrix of the metacells is renamed to fraction .

The __name__ scalar is not copied.

The excluded_gene and excluded_cell masks are not copied. Instead, if raw_cells_h5ad is specified, an is_excluded mask is created for both cells and genes, marking these that exist only in the raw_cells_h5ad and not in clean_cells_h5ad and metacells_h5ad .

The full_gene_index is not copied.

The properly_sampled_gene mask is renamed to the per-gene is_properly_sampled mask.

The bursty_lonely_gene mask is renamed to the per-gene is_bursty_lonely mask.

The lateral_gene mask is renamed to the per-gene is_lateral mask.

The noisy_gene mask is renamed to the per-gene is_noisy mask.

The rare_gene mask is renamed to the per-gene is_rare mask.

The rare_gene_module has 1 added to it (that is, "no module" is 0 in Daf ) and is renamed to rare_module .

The lateral_genes_module has 1 added to it (that is, "no module" is 0 in Daf ) and is renamed to lateral_module .

The marker_gene mask is renamed to the per-gene is_marker mask.

The selected_gene mask is renamed to the per-gene is_selected mask.

The ignored_gene mask is renamed to the per-gene is_ignored mask.

The ignored_gene_of_<type> masks are converted to an is_ignored mask per-gene-per-type.

The projected_noisy_gene mask is renamed to the per-gene is_projected_noisy mask.

The atlas_gene , atlas_lateral_gene , atlas_noisy_gene , atlas_marker_gene masks are renamed to the is_atlas , is_atlas_lateral , is_atlas_noisy and is_atlas_marker per-gene masks.

The essential_gene_of_<type> masks are converted to an is_essential mask per-gene-per-type.

The atlas_essential_gene_of_<type> masks are converted to an is_atlas_essential mask per-gene-per-type.

The fitted_gene_of_<type> masks are converted to an is_fitted mask per-gene-per-type.

The fitted mask per-gene-per-metacell is renamed to is_fitted .

The misfit mask per-gene-per-metacell is renamed to is_misfit .

The essential mask per-gene-per-metacell is renamed to is_essential .

The full_cell_index is not copied.

The properly_sampled_cell mask is renamed to the per-cell is_properly_sampled mask.

The rare_cell mask is renamed to the per-cell is_rare mask.

The cells_rare_gene_module has 1 added to it (that is, "no module" is 0 in Daf ) and is renamed to rare_gene_module .

The per-cell dissolve mask is renamed to is_dissolved .

The per-cell metacell integer annotation is not copied, and the metacell_name string annotation is renamed to metacell .

The per-cell most_similar integer annotation is not copied, and the most_similar_name string annotation is renamed to metacell.most_similar .

The rare_metacell mask is renamed to the per-metacell is_rare mask.

The per-metacell metacells_level is renamed to level .

The per-metacell similar mask is renamed to is_similar .

Note

There is much duplication of data between the three h5ad files (in particular, per-gene data). Data in raw_cells_h5ad will override data in clean_cells_h5ad , which will override data in metacells_h5ad .

Note

It is common to call reconstruct_axis! on the result (e.g., if the cells were collected from a set of batches).

Index

Metacells.AnnDataFormat

Metacells.AnnDataFormat.import_h5ads!