Compute Blocks

Metacells.ComputeBlocks — Module

Group "very close" metacells into disjoint blocks, and blocks into overlapping vicinities, for constructing a local linear approximation to the manifold.

Metacells.ComputeBlocks.compute_blocks! — Function

function compute_blocks!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = ```1.0e-5```,
    min_significant_gene_UMIs::Integer = ```40```,
    fold_confidence::AbstractFloat = ```0.9```,
    max_block_span::Real = ```2```,
    overwrite::Bool = ```false```,
)::Nothing

Group the metacells into blocks where the metacells in each one are "similar" in all the skeleton genes. That is, each block is an approximation of a single cell state, assuming the skeleton genes adequately predict the rest of the genes.

First, we compute metacell-metacell distances based on the maximal fold factor between is_skeleton genes. The fold factor is log (base 2) of the gene expression using the gene_fraction_regularization . For computing this fold factor, we ignore "insignificant" genes whose total UMIs in the compared metacells isn't at least min_significant_gene_UMIs . We also we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells. Two metacells can only belong to the same block if the final fold factor is at most max_block_span in all the (significant, skeleton) genes.

We then compute block-block distances which are the maximal distance between the metacells of the blocks.

Note

This uses the virtual metacell_gene_covered_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

metacell @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of the covered genes in each metacell.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): Some estimated fraction of the UMIs of each covered gene in each metacell.

gene, metacell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Outputs

Axes

block (guaranteed): Distinct groups of metacells with "very close" estimated cell state.

Vectors

metacell @ block ::AbstractString (guaranteed): The unique block each metacell belongs to.

Matrices

metacell, metacell @ max skeleton fold_distance ::Union{Float32, Float64} (guaranteed): The maximal fold factor between skeleton genes between the metacells.

block, block @ max skeleton fold_distance ::Union{Float32, Float64} (guaranteed): The maximal fold factor between skeleton genes between the metacells of the blocks.

Metacells.ComputeBlocks.compute_blocks_is_in_neighborhood! — Function

function compute_blocks_is_in_neighborhood!(
    daf::DafWriter;
    min_blocks_in_neighborhood::Integer = ```4```,
    min_metacells_in_neighborhood::Integer = ```20```,
    min_covered_UMIs_in_neighborhood::Integer = ```2000000```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Group the blocks into overlapping small tight neighborhoods, and larger environments, such that each block's environment can be reasonably approximated using a local linear model, evaluated using cross-validation on the RMSE in the neighborhood at its core.

Using the block-block distances, we define a tight neighborhood of each block containing at least min_blocks_in_neighborhood , min_metacells_in_neighborhood and min_covered_UMIs_in_neighborhood (note that the latter only counts UMIs of skeleton genes).

Having computed the neighborhoods, we expand them (using the same block-block distances) as long as computing a local linear model for the environment gives a better approximation of the metacells in the neighborhood. This local model starts with computing max_principal_components . These are assumed to over-fit the solution, so we use cross-validation (in cross_validation_parts ) to pick a subset of these principal components, which hopefully approximates the true dimensionality of the local data.

Note

There is no attempt to reconcile the local model of neighboring blocks. In general the linear model computed here isn't very useful for interpreting the data as principal components for noisy high-dimensional data are pretty opaque.

Inputs

Axes

block (required): Distinct groups of metacells with "very close" estimated cell state.

Vectors

block @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of the covered genes in each block.

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

Matrices

block, block @ max skeleton fold_distance ::Union{Float32, Float64} (required): The maximal fold factor between skeleton genes between the metacells of the blocks.

Outputs

Matrices

block, block @ is in neighborhood ::Bool (guaranteed): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

Metacells.ComputeBlocks.compute_blocks_is_in_environment! — Function

function compute_blocks_is_in_environment!(
    daf::DafWriter;
    max_principal_components = ```40```,
    cross_validation_parts::Integer = ```5```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Expand the block neighborhoods to larger linear environment. We expand the neighborhood as long as computing a local linear model for the environment gives a better approximation of the metacells in the neighborhood. This local model with computing max_principal_components . These are assumed to over-fit the solution, so we use cross-validation (in cross_validation_parts ) to pick a subset of these principal components to actually use, which hopefully approximates the true dimensionality of the local data.

Note

This uses the virtual metacell_gene_covered_fraction_matrix . You will need an adapter to map these to concrete fractions (geomean, linear, scaled, ...).

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

Vectors

gene @ is_covered ::Bool (required): A mask of genes that are covered by the local linear program.

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ covered_fraction ::Union{Float32, Float64} (required): Some estimated fraction of the UMIs of each covered gene in each metacell.

block, block @ is in neighborhood ::Bool (required): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

block, block @ max skeleton fold_distance ::Union{Float32, Float64} (required): The maximal fold factor between skeleton genes between the metacells of the blocks.

Outputs

Axes

principal_component (guaranteed): A local (per block) principal component.

Vectors

block @ n used principal_components ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): The number of used principal components for each block.

block @ pca_RMSE ::Union{Float32, Float64} (guaranteed): The root mean squared error of predicting the covered genes using the principal components in the vicinity of the block.

block @ pca_XRMSE ::Union{Float32, Float64} (guaranteed): The cross-validated root mean squared error of predicting the covered genes using the principal components in the vicinity of the block.

Matrices

block, block @ is in environment ::Bool (guaranteed): For each block, the (column) mask of nearby blocks in its expanded environment.

block, principal component @ is used ::Bool (guaranteed): Whether each principal component is used by each block.

block, gene @ base covered fraction ::Union{Float32, Float64} (guaranteed): The base covered fractions used by the principal component analysis..

Tensors

block; principal component, gene @ skeleton coefficient ::Union{Float32, Float64} (guaranteed): The coefficient of each skeleton gene for computing a principle component.

block; principal component, gene @ covered coefficient ::Union{Float32, Float64} (guaranteed): The coefficient of each principal component for computing each covered gene.

Index

Metacells.ComputeBlocks

Metacells.ComputeBlocks.compute_blocks!

Metacells.ComputeBlocks.compute_blocks_is_in_environment!

Metacells.ComputeBlocks.compute_blocks_is_in_neighborhood!