Compute Blocks

Metacells.ComputeBlocks — Module

Group "very close" metacells into disjoint blocks, and blocks into overlapping vicinities, for constructing a local linear approximation to the manifold.

Metacells.ComputeBlocks.compute_blocks! — Function

function compute_blocks!(
    daf::DafWriter;
    max_block_span::Real = ```2```,
    num_blocks::Integer = ```nothing```,
    overwrite::Bool = ```false```,
)::Nothing

Group the metacells into blocks where the metacells in each one are "similar" in all the skeleton genes. That is, each block is an approximation of a single cell state, assuming the skeleton genes adequately predict the rest of the genes.

TODOX

Inputs

Axes

metacell (required): Minimal-sized groups of cells for robust point estimates.

Matrices

metacell, metacell @ max skeleton fold_distance ::Union{Float32, Float64} (optional): The maximal significant fold factor between the covered fraction of skeleton genes between the metacells.

metacell, metacell @ euclidean skeleton distance ::Union{Float32, Float64} (optional): The Euclidean distance between the log of the covered fraction of the skeleton genes between the metacells.

Outputs

Axes

block (guaranteed): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (guaranteed): The unique block each metacell belongs to.

Metacells.ComputeBlocks.compute_blocks_is_in_neighborhood! — Function

function compute_blocks_is_in_neighborhood!(
    daf::DafWriter;
    min_blocks_in_neighborhood::Integer = ```5```,
    min_metacells_in_neighborhood::Integer = ```20```,
    min_covered_UMIs_in_neighborhood::Integer = ```2000000```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

TODOX

Group the blocks into overlapping small tight neighborhoods, and larger environments, such that each block's environment can be reasonably approximated using a local linear model, evaluated using cross-validation on the RMSE in the neighborhood at its core.

Using the block-block distances, we define a tight neighborhood of each block containing at least min_blocks_in_neighborhood , min_metacells_in_neighborhood and min_covered_UMIs_in_neighborhood (note that the latter only counts UMIs of skeleton genes).

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of the all covered genes in each block.

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

Matrices

block, block @ max skeleton fold_distance ::Union{Float32, Float64} (optional): The maximal significant skeleton genes covered fractions fold factor between metacells of the blocks.

block, block @ mean euclidean skeleton_distance ::Union{Float32, Float64} (optional): The mean Euclidean skeleton genes covered fractions distance between the metacells of the blocks.

Outputs

Matrices

block, block @ is in neighborhood ::Bool (guaranteed): For each block, the (column) mask of nearby blocks in its immediate neighborhood.

Metacells.ComputeBlocks.compute_blocks_is_in_environment! — Function

function compute_blocks_is_in_environment!(
    daf::DafWriter;
    overwrite::Bool = ```false```,
)::Nothing

TODOX

Group the blocks into overlapping small tight neighborhoods, and larger environments, such that each block's environment can be reasonably approximated using a local linear model, evaluated using cross-validation on the RMSE in the neighborhood at its core.

Using the block-block distances, we define a tight neighborhood of each block containing at least min_blocks_in_neighborhood , min_metacells_in_neighborhood and min_covered_UMIs_in_neighborhood (note that the latter only counts UMIs of skeleton genes).

Inputs

Axes

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

block @ covered_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of the all covered genes in each block.

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

Matrices

block, block @ max skeleton fold_distance ::Union{Float32, Float64} (optional): The maximal significant skeleton genes covered fractions fold factor between metacells of the blocks.

block, block @ mean euclidean skeleton_distance ::Union{Float32, Float64} (optional): The mean Euclidean skeleton genes covered fractions distance between the metacells of the blocks.

Outputs

Matrices

block, block @ is in environment ::Bool (guaranteed): For each block, the (column) mask of nearby blocks in its expanded environment.

Index

Metacells.ComputeBlocks

Metacells.ComputeBlocks.compute_blocks!

Metacells.ComputeBlocks.compute_blocks_is_in_environment!

Metacells.ComputeBlocks.compute_blocks_is_in_neighborhood!