Compute Modules

Metacells.ComputeModules — Module

Group covered genes in each environment to modules, each associated with some anchor gene.

Metacells.ComputeModules.compute_blocks_modules! — Function

function compute_blocks_modules!(
    daf::DafWriter;
    mean_genes_per_cluster::Integer = ```200```,
    anchor_correlation_genes::Integer = ```30```,
    min_anchor_correlation_genes_fraction::AbstractFloat = ```0.05```,
    max_anchor_correlation_genes_fraction::AbstractFloat = ```0.5```,
    min_gene_migration_correlation::AbstractFloat = ```0.7```,
    max_anchor_kept_cluster_genes_baseline::Integer = ```100```,
    min_anchor_kept_cluster_genes_likelihood::AbstractFloat = ```1.5```,
    overwrite::Bool = ```false```,
)::Nothing

Group covered genes in each environment to modules, each associated with some anchor gene.

For each block:

Compute the correlation of the expression of the covered genes across all metacells in the block's environment.

Cluster the genes based on this correlation, using Ward. The number of clusters is such that the mean number of genes per cluster is at least mean_genes_per_cluster .

For each cluster, pick an anchor gene out of the skeleton genes. For each skeleton, look at its worst correlation in the top anchor_correlation_genes (but at least min_anchor_correlation_genes_fraction and at most max_anchor_correlation_genes_fraction ). Pick as anchor the skeleton gene with the highest result.

For each gene in the cluster, keep it in the anchor's module if it is more correlated with the anchor than with any other anchor. Any gene with a correlation of more than min_gene_migration_correlation with some anchor is moved to the module of its most correlated anchor, regardless of the original cluster it was in.

For each cluster, consider the up to max_anchor_kept_cluster_genes_baseline genes most correlated with the cluster's anchor. If the fraction of the most correlated max_anchor_kept_cluster_genes_baseline genes that were kept in the anchor module is less than min_anchor_kept_cluster_genes_likelihood times the base fraction of 1/number of anchors, then discard the anchor.

The result is one module per anchor with the genes that were kept from its original cluster or were migrated to it. Some (most) genes will not participate in any module.

We name the resulting modules after their anchor gene for interpretability, but add a .MOD suffix to clarify this is the name of a module and not of a gene. It is possible that an anchor in one block will end up being a member of a different anchor's module in a neighbouring block, or even not to be used in any of the neighbouring block gene modules.

Due to Daf limitations, the modules axis must therefore be the union of all the modules from all the blocks, so we provide an is_found mask to specify which module is used for each block.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

Vectors

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

gene @ is_skeleton ::Bool (required): A mask of genes that are used to predict the values of the rest of the (covered) genes.

Matrices

block, block @ is in environment ::Bool (required): For each block, the (column) mask of nearby blocks in its expanded environment.

block, gene @ is environment marker ::Bool (required): A mask of covered genes that distinguish between cell states in each environment.

metacell, gene @ log covered fraction ::Union{Float32, Float64} (required): The log base 2 of the linear fraction of the UMIs of each covered gene in each metacell, out of the total covered UMIs.

Outputs

Axes

module (guaranteed): Groups of genes that together predict the rest of the covered genes.

Vectors

module @ gene.anchor ::AbstractString (guaranteed): The anchor gene each module is based on.

Matrices

block, module @ is_found ::Bool (guaranteed): A mask of the modules that were found for each block.

block, gene @ module ::AbstractString (guaranteed): The module each gene belongs to in each block.

Index