Contracts

Metacells.Contracts — Module

Functions for defining a Contract for a metacells Daf data set @computation . This also serves as a quick vocabulary of the contents of a metacells Daf data set.

In the descriptions below, "fold factor" refers to the log base 2 of the ratio between expression levels. For gene RNA expression levels, we typically use a regularization factor of 1e-5; fold factors of 1 (2x) is typically considered to be irrelevant, fold factors of 2 (4x) are considered to be potentially significant, fold factor of 3 (8x) are considered to point to a definite significant difference, etc. When used as thresholds, fold factors are adjusted using confidence levels (based on the number of UMIs used for the estimates) and the gene_divergence_vector .

Axes

Metacells.Contracts.gene_axis — Function

function gene_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of sequenced genes. By convention we use TODO as the namespace of the genes, but this may be different depending on the data set.

Metacells.Contracts.cell_axis — Function

function cell_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of sequenced single cells. There's no convention for cell names, as long as they are unique. Typically some sort of barcode is used, possibly combined with a batch and/or plate and/or experiment identification. In the latter case it is recommended that batch and/or plate and/or experiment would also be created as explicit axes, to allow associating metadata with them instead of repeating it for each cell.

Metacells.Contracts.metacell_axis — Function

function metacell_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of metacells, which are minimal-sized groups of cells for robust point estimates. That is, each metacell is considered to be a robustly estimated point in the multi-dimensional manifold of cell states. Metacells may be very similar or very different from each other depending on the data set and the manifold region.

Metacells.Contracts.box_axis — Function

function box_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of boxes, which are distinct groups of metacells with "very close" estimated cell state. That is, for some chosen set of genes, the metacells in each box all have very close estimates of gene expressions (maximal fold factor up to some maximal span). This maximal span is small and identical for all the boxes.

Metacells.Contracts.neighborhood_axis — Function

function neighborhood_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of neighborhoods, which are overlapping groups of boxes with "close" estimated cell states. That is, for some chosen set of genes, the metacells in all the boxes of the neighborhood will all have close estimates of gene expression (maximal fold factor up to some maximal span). This maximal span is moderate and different for each neighborhood to ensure neighborhood sizes are not too big.

Metacells.Contracts.type_axis — Function

function type_axis(expectation::ContractExpectation)::Pair{AxisKey, AxisSpecification}

The axis of types, which are distinct named biological cell states. That is, types are convenient labels manually assigned to large groups of cells, metacells, boxes, neighborhoods, etc. The resolution of the type labels depends on the data set and the type of analysis. In particular, types are not typically associated with a specific biological cell state, but rather with a set of related biological cell states (possibly along a gradient of such states).

Vectors

Metacells.Contracts.gene_is_excluded_vector — Function

function gene_is_excluded_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A mask of genes that are excluded from consideration. These genes are completely unrelated to the biological behaviors of interest. Not only that, they have strong and variable expression levels; enough to have an global impact on the expression level of the rest of the genes - for example, mitochondrial genes. Such genes make it difficult to estimate the relative expression level of genes between different cell states. Therefore, such genes aren't even counted in the total UMIs of each cell.

Metacells.Contracts.gene_is_lateral_vector — Function

function gene_is_lateral_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A mask of genes that are lateral to the biological behaviors of interest. That is, these genes may satisfy all criteria for being in a group of cooperating genes, but the biological behavior they participate in isn't relevant to the behaviors of interest - for example, genes related to cell cycle or stress. Such genes make it harder to focus on the biological behaviors of interest. They are therefore masked out during the analysis.

Metacells.Contracts.gene_is_marker_vector — Function

function gene_is_marker_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A mask of genes that distinguish between cell states. That is, these genes have a significant expression level at some cell state, as well as a significant range of expression across all cell states, so can be used to distinguish between cell states. Non-marker genes are by definition not useful for such analysis, but marker genes aren't necessarily useful due to other considerations.

Metacells.Contracts.gene_is_correlated_vector — Function

function gene_is_correlated_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A mask of genes that are correlated with other gene(s). We typically search for groups of genes that act together. Genes that have no correlation with other genes aren't useful for this sort of analysis, even if they are marker genes.

Metacells.Contracts.gene_divergence_vector — Function

function gene_divergence_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

For each gene, we scale fold factors of each gene by multiplying with (1 - divergence) of the gene. In particular this is used when considering the distance between gene expressions. Therefore genes with a non-zero divergence will require a higher fold factor

Metacells.Contracts.cell_is_excluded_vector — Function

function cell_is_excluded_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A mask of cells that are excluded from consideration. This can be due to any number of reasons - doublets, too low a number of UMIs, to high a percentage of excluded gene UMIs, etc.

Metacells.Contracts.metacell_total_UMIs_vector — Function

function metacell_total_UMIs_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The total number of UMIs used to estimate the fraction of all the genes in each metacell. This is used to estimate the robustness of the estimates.

Metacells.Contracts.metacell_box_vector — Function

function metacell_box_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The unique box each metacell belongs to.

Metacells.Contracts.metacell_type_vector — Function

function metacell_type_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The type each metacell belongs to.

Metacells.Contracts.box_main_neighborhood_vector — Function

function box_main_neighborhood_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The unique main neighborhood of each box. Ideally, each box is the center of its own main neighborhood, and also belongs to overlapping neighborhood of some other nearby boxes. However, where the manifold is sparsely sampled, a few nearby boxes may share the same main neighborhood. If the samples are sufficiently sparse, the main neighborhood may include only just the single box (which itself may include just a single metacell).

Metacells.Contracts.box_total_UMIs_vector — Function

function box_total_UMIs_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The unique main neighborhood of each box. Ideally, each box is the center of its own main neighborhood, and also belongs to overlapping neighborhood of some other nearby boxes. However, where the manifold is sparsely sampled, a few nearby boxes may share the same main neighborhood. If the samples are sufficiently sparse, the main neighborhood may include only just the single box (which itself may include just a single metacell).

Metacells.Contracts.box_type_vector — Function

function box_type_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The type each box belongs to.

Metacells.Contracts.neighborhood_span_vector — Function

function neighborhood_span_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

The span (fold factor) used to compute the neighborhood. This is different for each neighborhood (but is never too big), to ensure that the sizes of the neighborhoods are not too large even for densely sampled regions of the manifold (a simple fixed span will not do, due to the curse of multi-dimensionality).

Metacells.Contracts.type_color_vector — Function

function type_color_vector(expectation::ContractExpectation)::Pair{VectorKey, DataSpecification}

A unique color for each type for graphs.

Matrices

Metacells.Contracts.gene_metacell_fraction_matrix — Function

function gene_metacell_fraction_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

The estimated fraction of the UMIs of each gene in each metacell. We assume that each metacell is a sample of the manifold, representing a real biological state, regardless to its distance to other metacells (subject to cleaning up batch effects, purging doublets, and compensating for any other technical artifacts).

Metacells.Contracts.gene_metacell_total_UMIs_matrix — Function

function gene_metacell_total_UMIs_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

The total number of UMIs used to estimate the fraction of each gene in each metacell. This is used to estimate the robustness of the estimate. When computing fold factors, we require the total number of UMIs (from both compared estimates) to be some minimum, and possibly adjust the fold factor according to some confidence level (assuming a multinomial sampling distribution).

Metacells.Contracts.gene_neighborhood_is_correlated_matrix — Function

function gene_neighborhood_is_correlated_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

Which genes are correlated in each neighborhood. This is a smaller set than the set of globally correlated genes, since some genes may be a strong marker for some cell type (and thus have a strong correlation with other genes specific to this type), but may lack correlation with any other genes when considering only cell states of this type. As such, they hamper analysis of groups of cooperating genes within this specific cell type. However, instead of relying on manual type annotations, we make use of the computed neighborhoods to obtain the set of correlated genes in each local region of the manifold.

Metacells.Contracts.box_box_distance — Function

function box_box_distance(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

The distance (fold factor) between the most different metacell genes between the boxes. This is the fold factor between the most different gene expression in a pair of metacells, one in each box. This considers only the chosen genes (marker genes that are also correlated in the main neighborhood of either of the boxes).

Metacells.Contracts.gene_box_fraction_matrix — Function

function gene_box_fraction_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

The estimated fraction of the UMIs of each gene in each box. Each box is a sample of the manifold, representing a real biological state, which is different from the state of any other box.

Metacells.Contracts.gene_box_total_UMIs_matrix — Function

function gene_box_total_UMIs_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

The total number of UMIs used to estimate the fraction of each gene in each box. This is used to estimate the robustness of the estimate. When computing fold factors, we require the total number of UMIs (from both compared estimates) to be some minimum, and possibly adjust the fold factor according to some confidence level (assuming a multinomial sampling distribution).

Metacells.Contracts.box_neighborhood_is_member_matrix — Function

function box_neighborhood_is_member_matrix(expectation::ContractExpectation)::Pair{MatrixKey, DataSpecification}

A mask of the member boxes of each neighborhood. This is needed since each box may belong to multiple neighborhoods, and each neighborhood contains multiple boxes.

Index

Metacells.Contracts

Metacells.Contracts.box_axis

Metacells.Contracts.box_box_distance

Metacells.Contracts.box_main_neighborhood_vector

Metacells.Contracts.box_neighborhood_is_member_matrix

Metacells.Contracts.box_total_UMIs_vector

Metacells.Contracts.box_type_vector

Metacells.Contracts.cell_axis

Metacells.Contracts.cell_is_excluded_vector

Metacells.Contracts.gene_axis

Metacells.Contracts.gene_box_fraction_matrix

Metacells.Contracts.gene_box_total_UMIs_matrix

Metacells.Contracts.gene_divergence_vector

Metacells.Contracts.gene_is_correlated_vector

Metacells.Contracts.gene_is_excluded_vector

Metacells.Contracts.gene_is_lateral_vector

Metacells.Contracts.gene_is_marker_vector

Metacells.Contracts.gene_metacell_fraction_matrix

Metacells.Contracts.gene_metacell_total_UMIs_matrix

Metacells.Contracts.gene_neighborhood_is_correlated_matrix

Metacells.Contracts.metacell_axis

Metacells.Contracts.metacell_box_vector

Metacells.Contracts.metacell_total_UMIs_vector

Metacells.Contracts.metacell_type_vector

Metacells.Contracts.neighborhood_axis

Metacells.Contracts.neighborhood_span_vector

Metacells.Contracts.type_axis

Metacells.Contracts.type_color_vector