Identify Genes

Metacells.IdentifyGenes — Module

Identify special genes.

Metacells.IdentifyGenes.compute_genes_divergence! — Function

function compute_genes_divergence!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1e-5,
    min_divergent_gene_range_fold::AbstractFloat = 6.0,
    overwrite::Bool = false,
)::Nothing

Compute a divergence factor for all genes. This factor is typically 0.0 for most genes (and is always below 1.0). Genes that have a wide range of expression are given a higher divergence factor. When looking for significant differences in gene expressions, the divergence factor is used to scale down the difference, to compensate for the gene's inherent wider expression range. Specifically, we take the raw fold factor and multiply it by (1.0 - divergence).

This works as follows:

Compute the minimal and maximal expression level of each gene.

Compute the fold factor (log2 of maximal over minimal value, using the gene_fraction_regularization .

The raw factor of the gene is the ratio between the fold factor and the min_divergent_gene_range_fold . If this is higher than 1, we set the divergence factor accordingly.

Note

Ideally, all code that uses the manual is_noisy mask should be modified to use the divergence factor instead.

CONTRACT

Metacells.IdentifyGenes.identify_marker_genes! — Function

function identify_marker_genes!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1e-5,
    min_gene_range_fold::AbstractFloat = 2.0,
    min_max_marker_gene_fraction::AbstractFloat = 1e-4,
    overwrite::Bool = false,
)::Nothing

Identify the genes that distinguish at least one metacell from the rest. Such genes are called "marker" genes as they (potentially) mark specific cell states. If overwrite , will overwrite an existing is_marker mask.

Compute the minimal and maximal expression level of each gene.

Compute the fold factor (log2 of maximal over minimal value, using the gene_fraction_regularization .

Reduce this fold factor using the divergence factor of the genes.

Identify as markers genes whose adjusted fold factor is at least min_marker_gene_range_fold , and whose maximal expression is at least min_max_marker_gene_fraction .

CONTRACT

Metacells.IdentifyGenes.identify_correlated_genes! — Function

function identify_correlated_genes!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 1e-5,
    min_gene_correlation::AbstractFloat = 0.5,
    overwrite::Bool = false,
)::Nothing

Identify genes that are correlated with other gene(s). Such genes are good candidates for looking for groups of genes that act together. If overwrite , will overwrite an existing is_correlated mask.

Compute the log base 2 of the genes expression in each metacell (using the gene_fraction_regularization ).

Correlate this between all the pairs of genes.

Find the maximal absolute correlation for each gene (that is, strong anti-correlation also counts).

Identify the genes which have at least one gene with a correlation of at least min_gene_correlation .

CONTRACT

Index

Metacells.IdentifyGenes

Metacells.IdentifyGenes.compute_genes_divergence!

Metacells.IdentifyGenes.identify_correlated_genes!

Metacells.IdentifyGenes.identify_marker_genes!