Programs

Metacells.Programs — Module

Approximate the manifold of actual cell states (captured by metacells) using linear programs in each local region.

Metacells.Programs.compute_factor_priority_of_genes! — Function

function compute_factor_priority_of_genes!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    max_principal_components::Integer = 40,
    factors_per_principal_component::Real = 2,
)::Nothing

Given the transcription factors in the data, identify an ordered subset of these to use as candidates for predicting the rest of the genes. To achieve this, we run PCA analysis with up to max_principal_components on the log (base 2) of the gene expressions (using the gene_fraction_regularization ). We then rank each gene in each principal component (based on the absolute coefficient value) and compute the minimal rank of each gene (giving lower weight to genes in later components). We prioritize the transcription factors according to this minimal rank, choosing the top factors_per_principal_component times the number of principal components.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is transcription factor ::Bool (required): A mask of genes that bind to the DNA.

gene @ is forbidden factor ::Bool (optional): A mask of genes that are forbidden from being used as predictive factors.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

gene @ factor_priority ::Union{UInt16, UInt32, UInt64, UInt8} (guaranteed): Priority for using the gene as a predictor for the rest of the genes.

Metacells.Programs.compute_global_predictive_factors! — Function

function compute_global_predictive_factors!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    cross_validation::Integer = 5,
    rng::AbstractRNG = default_rng(),
)::Nothing

Given a prioritized subset of the factor genes likely to be predictive, identify a subset of these genes which best predict the values of the rest of the genes, using non-negative least-squares approximation using the log (base 2) of the genes expression (using the gene_fraction_regularization ). To avoid over-fitting, we use cross_validation .

We require each additional used factor to improve (reduce) the RMS of the cross validation error by at least min_rms_improvement (on average). Factors with higher prioriry are allowed a lower improvement and factors with a lower priority require higher improvement; the range of this is rms_priority_improvement times the min_rms_improvement . That is, if this range is 1, then the highest priority factor requires 0.5 times the minimal improvement, and the lowest priority factor requires 1.5 times the minimal improvement.

Since this prediction is global for the whole data set, we do not expect it to be useful as of itself. We therefore intentionally do not store the coefficients. Instead we only store the mask of chosen predictive factors.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ factor_priority ::Union{UInt16, UInt32, UInt64, UInt8} (required): Priority for using the gene as a predictor for the rest of the genes.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

Outputs

Vectors

gene @ is global predictive_factor ::Bool (guaranteed): A mask of globally predictive transcription factors.

Metacells.Programs.compute_blocks! — Function

function compute_blocks!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    fold_confidence::AbstractFloat = 0.9,
    max_block_span::Real = 2,
)::Nothing

Given a set of transcription factors that can be used to predict the rest of the genes across the whole manifold, group the metacells into distinct blocks, such that within each block, the expression level of all these factors differ by a fold factor of more than max_block_span between any of the metacells of the block. When evaluating the difference in expression between the genes, we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells; we also ignore any difference between the expression level of genes whose total UMIs (in both compared metacells) is less thabn min_significant_gene_UMIs .

We expect all the metacells within each block to be "essentially identical", so blocks serve as a useful way to group metacells together for the purpose of type annotations. We also assume that the same local linear program applies for all metacells of each block. The distance matrix between the blocks expresses the overall structure of the manifold.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is global predictive_factor ::Bool (required): A mask of globally predictive transcription factors.

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Outputs

Axes

block (guaranteed): Distinct groups of metacells with "very close" estimated cell state.

Vectors

metacell @ block ::AbstractString (guaranteed): The unique block each metacell belongs to.

Matrices

block, block @ distance ::Union{Float32, Float64} (guaranteed): The mean distance (maximal fold factor of any gene) between the metacells of the blocks.

Metacells.Programs.compute_blocks_vicinities! — Function

function compute_blocks_vicinities!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    min_rms_improvement::AbstractFloat = 0.02,
    rms_priority_improvement::Real = 1,
    block_rms_bonus::AbstractFloat = 0.0005,
    cross_validation::Integer = 5,
    min_blocks_in_neighborhood::Integer = 4,
    min_metacells_in_neighborhood::Integer = 100,
    rng::AbstractRNG = default_rng(),
)::Nothing

Given a partition of the metacells into distinct blocks, compute for each block its immediate neighborhood (of at least min_blocks_in_neighborhood blocks and at least min_metacells_in_neighborhood metacells). Then expact this to an environment such that computing a linear model for the environment (based on the precomputed set of global predictive genes) minimizes the RMS of the error in the neighborhood.

The motivation is that the small neighborhood is large enough to evaluate the linear model, but is too small by itself for detecting the linear model. Expanding the neighborhood to a larger environment allows us to detect weak linear programs that exist in the neighborhood, therefore reducing the RMS. Since the manifold is not linear, expanding the environment beyond a certain point starts to increase the RMS again. We give a slight bonus of block_rms_bonus for per each block added to the environment to reduce the chances of stopping in a local minimum.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" estimated cell state.

Vectors

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is global predictive_factor ::Bool (required): A mask of globally predictive transcription factors.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

block, block @ distance ::Union{Float32, Float64} (required): The mean distance (maximal fold factor of any gene) between the metacells of the blocks.

Outputs

Matrices

block, block @ is in neighborhood ::Bool (guaranteed): For each block, the mask of nearby blocks in its immediate neighborhood.

block, block @ is in environment ::Bool (guaranteed): For each block, the mask of nearby blocks in its linear environment.

Metacells.Programs.compute_local_predictive_factors! — Function

function compute_local_predictive_factors!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    min_rms_improvement::AbstractFloat = 0.02,
    rms_priority_improvement::Real = 1,
    cross_validation::Integer = 5,
    min_marker_gene_range_fold::Real = 1,
    min_marker_gene_max_fraction::AbstractFloat = 0.0001,
    rng::AbstractRNG = default_rng(),
)::Nothing

Having computed the neighborhoods and environments, then for each block, figure out the set of transcription factors for best approximating the gene expression of the metacells in each neighborhood based on the environment. This set of local predictive factors will be different from the set of global predictive factors (and will be different for each block).

When computing this set, we only consider the genes that are marker genes within the environment (as per identify_marker_genes! , using min_marker_gene_max_fraction and a tighter min_marker_gene_range_fold .

Index

Metacells.Programs

Metacells.Programs.compute_blocks!

Metacells.Programs.compute_blocks_vicinities!

Metacells.Programs.compute_factor_priority_of_genes!

Metacells.Programs.compute_global_predictive_factors!

Metacells.Programs.compute_local_predictive_factors!