Programs

Metacells.Programs — Module

Approximate the manifold of actual cell states (captured by metacells) using linear programs in each local region.

Metacells.Programs.compute_blocks! — Function

function compute_blocks!(
    daf::DafWriter;
    gene_fraction_regularization::AbstractFloat = 2.0e-5,
    min_significant_gene_UMIs::Integer = 40,
    fold_confidence::AbstractFloat = 0.9,
    max_block_span::Real = 1,
    max_principal_components = 40,
    min_blocks_in_neighborhood::Integer = 4,
    min_metacells_in_neighborhood::Integer = 50,
    min_blocks_in_approximate_environment::Integer = 20,
    min_metacells_in_approximate_environment::Integer = 500,
    cross_validation_parts::Integer = 5,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = false,
)::Nothing

Group the metacells into blocks where the metacells in each one are "similar" in all the transcription factors. That is, each block is an approximation of a single cell state, assuming the transcription factors control the expression of the rest of the genes. Then, group these blocks into overlapping small tight neighborhoods, and larger environments, such that each block's environment can be reasonably approximated using a linear model (which is not computed), evaluated using cross-validation on the RMSE in the neighborhood at its core.

First, we compute metacell-metacell distances based on the maximal fold factor between is_transcription_factor genes, which aren't is_forbidden_factor . The fold factor is log (base 2) of the gene expression using the gene_fraction_regularization . For computing this fold factor, we ignore genes whose total UMIs in the compared metacells isn't at least min_significant_gene_UMIs . We also we reduce the distance using the fold_confidence based on the number of UMIs used to estimate the expression in the metacells, and the divergence of the gene. Two metacells can only belong to the same block if the final fold factor is at most max_block_span in all the genes.

We then compute block-block distance which is the mean distance between the metacells of the blocks. Using this distances, we define a tight neighborhood of each block containing at least min_blocks_in_neighborhood and min_metacells_in_neighborhood . We also compute an approximate environment of at least min_blocks_in_approximate_environment and min_metacells_in_approximate_environment . Using this approximate environment, we compute max_principal_components for the transcription factor genes, and figure out how many of them are useful to predict the value of the rest of the genes using cross_validation_parts minimizing the RMSE in the neighborhood.

Having determined the dimensionality of the environment of the block, we search for the optimal environment such that the RMSE of the cross validation of a linear model with this humber of principal components (computed on the transcription factor genes) for predicting the value of the rest of the genes.

Inputs

Axes

gene (required): Sequenced genes.

metacell (required): Minimal-sized groups of cells for robust point estimates.

Vectors

metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of all the genes in each metacell.

gene @ divergence ::Union{Float32, Float64} (required): Scale fold factors of each gene by multiplying with (1 - divergence) of the gene.

gene @ is transcription factor ::Bool (required): A mask of genes that bind to the DNA.

gene @ is forbidden factor ::Bool (optional): A mask of genes that are forbidden from being used as predictive factors.

Matrices

gene, metacell @ fraction ::Union{Float32, Float64} (required): The estimated fraction of the UMIs of each gene in each metacell.

gene, metacell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs used to estimate the fraction of each gene in each metacell.

Additional inputs may be used depending on the parameter(s).

Outputs

Axes

block (guaranteed): Distinct groups of metacells with "very close" estimated cell state.

Vectors

metacell @ block ::AbstractString (guaranteed): The unique block each metacell belongs to.

Matrices

block, block @ is in neighborhood ::Bool (guaranteed): For each block, the mask of nearby blocks in its immediate neighborhood.

block, block @ is in environment ::Bool (guaranteed): For each block, the mask of nearby blocks in its linear environment.

Additional outputs may be created depending on the parameter(s).

Index

Metacells.Programs

Metacells.Programs.compute_blocks!