Sharpen Metacells

Metacells.SharpenMetacells — Module

Compute "better" metacells based on the local linear model approximating the manifold.

Metacells.SharpenMetacells.sharpen_metacells! — Function

function sharpen_metacells!(;
    sharp_daf::DafWriter,
    base_daf::DafReader,
    prefix::AbstractString = ```"M"```,
    min_cells_in_metacell::Integer = ```12```,
    min_migration_likelihood::AbstractFloat = ```1.5```,
    kmeans_rounds::Integer = ```10```,
    rng::AbstractRNG = default_rng(),
    overwrite::Bool = ```false```,
)::Nothing

Given an base_daf metacells repository with a blocks structure and local gene modules that describe the cell state manifold, compute a sharp_daf metacells repository, which hopefully more faithfully captures this manifold.

We cluster using K-means all the cells in each neighborhood using the z-score of the expression of the modules of that neighborhood. The number of clusters is the number of base metacells in that neighborhood.

We assign to each cluster the base block which is most frequent in the cluster cells. Cells of the base base block of the neighborhood, which belong to a cluster that is assigned to a different block, and which also belong to a cluster of that block in the neighborhood of that block, are migrated to that block, but only if the enrichment of the cells of that block in the cluster is at least min_migration_likelihood times what would be expected assuming random clustering based on the relative sizes of the base and other blocks.

Having finalized the block to which each cell belongs to, we cluster all the cells in each block using K-means using the modules of the neighborhood of that block. We start with the expected number of metacells in that block (based on the mean number of cells per metacell in the base block) and adjust the number of clusters to try and enforce the sizes of the clusters - not more than twice that mean and no less than min_cells_in_metacell . In edge cases we dissolve too-small clusters, so this can create new outlier cells.

The final clusters are the sharp metacells. We name them using the prefix , the convention is to advance the letter for each sharpening round ( M to N to O to ...).

Whenever we call K-means we repeat the call kmeans_rounds times and pick the best result.

Sharpened Metacells

Inputs

Axes

cell (required): Sequenced single cells.

Outputs

Axes

metacell (created): Minimal-sized groups of cells for robust point estimates.

Vectors

cell @ metacell ::AbstractString (created): The unique metacell each cell belongs to.

metacell @ base_block ::AbstractString (created): The unique base block each sharpened metacell belongs to.

Original Metacells

Inputs

Axes

cell (required): Sequenced single cells.

metacell (required): Minimal-sized groups of cells for robust point estimates.

block (required): Distinct groups of metacells with "very close" cell state.

gene (required): Sequenced genes.

module (required): Groups of non-lateral genes that together predict the rest of the genes at a specific region of the manifold.

Vectors

cell @ total_UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of UMIs of all the non-excluded genes in each cell.

cell @ metacell ::AbstractString (required): The unique metacell each cell belongs to.

metacell @ block ::AbstractString (required): The unique block each metacell belongs to.

block @ n_cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of cells in each block.

block @ n_metacells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of metacells in each block.

block @ n_modules ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of found modules in each block.

block @ n neighborhood cells ::Union{UInt16, UInt32, UInt64, UInt8} (required): The total number of cells in the metacells in each block's neighborhood.

Matrices

gene, cell @ UMIs ::Union{UInt16, UInt32, UInt64, UInt8} (required): The number of UMIs collected for each gene for each cell.

block, gene @ module ::AbstractString (required): The module each gene belongs to in each block.

block, module @ is_found ::Bool (required): A mask of the modules that were found for each block.

block, block @ is in neighborhood ::Bool (required): For each (column) block, whether each (row) block is in its immediate neighborhood.

block, module @ mean linear fraction in neighborhood_cells ::Union{Float32, Float64} (required): The mean of the linear fraction of each module in the cells of the neighborhood of each block.

block, module @ std linear fraction in neighborhood_cells ::Union{Float32, Float64} (required): The standard deviation of the linear fraction of each module in the cells of the neighborhood of each block.

Index

Metacells.SharpenMetacells

Metacells.SharpenMetacells.sharpen_metacells!