Spheres

Metacells.Spheres — Module

Given a set of raw metacells, partition them into spheres such that all metacells in the same sphere are within some (fold factor) radius of each other. The centroids of these spheres can serve as a representation of the cell state manifold which is less sensitive to oversampling of common cell states. Group these spheres in overlapping neighborhoods of "similar" spheres for further analysis.

Metacells.Spheres.compute_spheres! — Function

function compute_spheres!(
    daf::DafWriter;
    min_significant_gene_UMIs::Integer = 40,
    gene_fraction_regularization::AbstractFloat = 1e-5,
    confidence::AbstractFloat = 0.9,
    max_sphere_diameter::AbstractFloat = 2.0,
    max_neighborhood_diameter::AbstractFloat = 2.0,
    min_gene_correlation::AbstractFloat = 0.5,
    max_deviant_genes_fraction::AbstractFloat = 0.01,
    overwrite::Bool = false,
)::Nothing

Partition raw metacells into distinct spheres, and spheres into overlapping neighborhoods.

Initial spheres and neighborhoods are computed in a first round, and then refined in a series of followup rounds.

In each round, we compute a distance between each two metacells. This is based on the fold factor between the expression level of each (relevant) gene in the metacells. The fold factor is the absolute value of the difference in the log (base 2) of the fraction of the gene in the metacells. This log is computed with the gene_fraction_regularization (by default, 1e-5 ). Since the fraction of the gene is a random variable, we decrease the high fraction and increase the low fraction by a factor based on the confidence of the test (by default, 0.9), assuming a multinomial distribution. In addition, if the sum of the total UMIs of the gene in both metacells is less than min_significant_gene_UMIs (by default, 40 ), we ignore this fold factor as insignificant. Finally, genes, we reduce the fold factor using the gene's divergence factor. In the first round, we simply count the number of genes whose fold factor is more than max_sphere_diameter (for computing spheres) and max_sphere_diameter + max_neighborhood_diameter (for computing neighborhoods). In the followup rounds, we use the maximal gene fold, for genes that are correlated in the vicinity of the metacells (see below). Finally, in the followup rounds, we only compare a base metacell with other metacells which were in the same neighborhood in the previous round.

We use hierarchical clustering to partition the metacells to distinct spheres, such that the maximal distance between any metacells in the sphere is bounded. In the first round, this bound is the max_deviant_genes_fraction out of the total number of genes. In the followup rounds, this is the max_sphere_diameter .

For each sphere, we compute a main neighborhood of other spheres such that the maximal distance between any metacells in the neighborhood is bounded. In the first round, this bound is again the maximal number of deviant genes (this time, using the increased fold distance computed above). In the followup rounds, this is the max_sphere_diameter plus the max_neighborhood_diameter . These neighborhoods may overlap. The main neighborhoods of different spheres may even be identical.

For each sphere, we compute the set of genes which have at least the min_gene_correlation with some other gene(s) in its main neighborhood. We restrict the correlated set of genes of each metacell to be the intersection of this set with the set from its sphere in the previous round.

If the new sets of correlated genes only differ up to max_convergence_fraction , consider this to have converged. Otherwise, repeat from step 2.

If overwrite is set, the results will replace any previously computed spheres and neighborhoods.

CONTRACT

Index

Metacells.Spheres

Metacells.Spheres.compute_spheres!